### 2020-05-21 #2197 @whedon
[PRE REVIEW]: IPART: A Python Package for Image-Processing based Atmospheric River Tracking
**Submitting author:** @Xunius (<a href="http://orcid.org/0000-0001-5877-5608">Guangzhi XU</a>)
**Repository:** <a href="https://github.com/ihesp/IPART" target ="_blank">https://github.com/ihesp/IPART</a>
**Version:** v1.0
