### 2020-05-21 #2197 @kthyng
@openjournals/dev Is there a reference for finding what characters are not allowed in the papers? It looks like that is the problem. Is & ok?
