### 2020-05-21 #2197 @danielskatz
> @openjournals/dev Is there a reference for finding what characters are not allowed in the papers? It looks like that is the problem. Is & ok?

This is not an issue with the paper body, but with the header.
Specifically, on line 9
	orcid: 0000-0001-5877-5608
should be
    orcid: 0000-0001-5877-5608

(basically, a tab needs to be replaced with 4 spaces)
