### 2020-05-21 #2197 @Xunius
Thanks. Just made a quick fix: https://github.com/ihesp/AR_tracker/commit/6e617d8f0d5ad43beec1dbbf8ea693f4e57013cf
