### 2020-05-26 8077a72
fix np.set_printoptions()
river_tracker1.py, river_tracker2.py, 3_detect_ARs.ipynb:

    use `np.set_printoptions(threshold=np.inf)` for py2,
    `np.set_printoptions(threshold=sys.maxsize)` for py3. Not sure
    why can't use the same for both versions.
