### 2020-05-27 #2197 @Xunius
@kbarnhart I noticed that generated PDF is using this format of citation `[@authoryear]` with no reference list in the end. In the PDF I compiled myself (using the preview tool or a Llinux commandline utility, I can't remember which exactly) the citations look correct. Is it something again related to formatting errors?
