### 2020-05-27 16dc616
enhance inland penetration of detected ARs
compute_thr_singlefile.py:

    add ORO_FILE and HIGH_TERRAIN global parameters.

    add optional `oro` and `high_terrain` input args to filterData().

    in filterData(), use a 6-connectivity structuring element for 3D
    reconstruction.

    in filterData(), if `oro` is given, perform an extra reconstruction
    over regions of high terrain. And merge results to global THR with a
    elementwise maximum. If `oro` is None, compute THR as usual.

    Doc changes.

compute_thr_multifile.py:

    add ORO_FILE and HIGH_TERRAIN global parameters.

    
[…]
