+++ river_tracker1.py
@@ -49,23 +49,24 @@ from __future__ import print_function
 #######################################################################
 
 #--------------------Time range--------------------
-YEAR=1984
+YEAR=2007
 TIME_START='%d-01-01 00:00:00' %YEAR
-TIME_END='%d-06-30 18:00:00' %YEAR
+TIME_END='%d-12-31 18:00:00' %YEAR
 
 #-----------u-qflux----------------------
-SOURCEDIR1='/home/guangzhi/datasets/erai/ERAI_AR_THR'
-UQ_FILE_NAME='uflux_m1-60_6_%d_crop.nc'
+SOURCEDIR1='/home/guangzhi/datasets/erai_qflux/'
+UQ_FILE_NAME='uflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
 UQ_VAR='uflux'
 
 #-----------v-qflux----------------------
-SOURCEDIR2='/home/guangzhi/datasets/erai/ERAI_AR_THR'
-VQ_FILE_NAME='vflux_m1-60_6_%d_crop.nc'
+SOURCEDIR2='/home/guangzhi/datasets/erai_qflux'
+VQ_FILE_NAME='vflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
 VQ_VAR='vflux'
 
 #-----------------ivt reconstruction and anomalies-----------------
 SOURCEDIR3='/home/guangzhi/datasets/erai/ERAI_AR_THR/'
-IVT_FILE_NAME='ivt_m1-60_6_%d_crop-minimal-rec-ano-kernel-t16-s6.nc'
+IVT_FILE_NAME='ivt_m1-60_6_%d_cln-cea-proj-THR-kernel-t16-s6.nc' %YEAR
+
 
 #------------------Output folder------------------
 OUTPUTDIR='/home/guangzhi/datasets/erai/ERAI_AR_THR/%d/' %YEAR
@@ -73,7 +74,6 @@ OUTPUTDIR='/home/guangzhi/datasets/erai/ERAI_AR_THR/%d/' %YEAR
 
 
 PLOT=True          # create maps of found ARs or not
-SINGLE_DOME=False  # do peak partition or not
 
 LAT1=0; LAT2=90      # degree, latitude domain
 # NOTE: this has to match the domain seletion in compute_thr_singlefile.py
@@ -105,8 +105,11 @@ PARAM_DICT={
     'rdp_thres': 2,
     # grids. Remove small holes in AR contour.
     'fill_radius': max(1,int(4*0.75/RESO)),
-    # max prominence/height ratio of a local peak. Only used when SINGLE_DOME=True
-    'max_ph_ratio': 0.4,
+    # do peak partition or not, used to separate systems that are merged
+    # together with an outer contour.
+    'single_dome': False,
+    # max prominence/height ratio of a local peak. Only used when single_dome=True
+    'max_ph_ratio': 0.6,
     # minimal proportion of flux component in a direction to total flux to
     # allow edge building in that direction
     'edge_eps': 0.4
@@ -135,37 +138,22 @@ from river_tracker1_funcs import areaFilt, maskToGraph, getARAxis, cropMask,\
 
 
 
-def findARs(anoslab, quslab, qvslab, areas, costhetas, sinthetas, param_dict):
+def findARs(anoslab, areas, param_dict):
     '''Find ARs from a time snap
 
     Args:
         anoslab (cdms.TransientVariable): (n * m) 2D anomalous IVT slab, in kg/m/s.
-        quslab (cdms.TransientVariable): (n * m) 2D u-flux slab, in kg/m/s.
-        qvslab (cdms.TransientVariable): (n * m) 2D v-flux slab, in kg/m/s.
[…]
