### 2020-05-27 570df92
fix citations in joss
