+++ joss/paper.md
@@ -34,15 +34,15 @@ term was coined as an analogy to the terrestrial rivers in a sense
 that when viewed from satellite imagery or large scale atmospheric
 observation, they appear as narrow and elongated vapor filaments,
 representing transient intensified horizontal moisture fluxes
-(e.g. `@Gimeno2014`, `@Dettinger2011`). A typical atmospheric river
+(e.g. @Gimeno2014, @Dettinger2011). A typical atmospheric river
 can carry 7-15 times the water in the Mississippi River
-`[@Ralph2011]`, and at any time in winter, there are four to five such
-systems in the Northern Hemisphere alone `[@Zhu1998]`, accounting for
+[@Ralph2011], and at any time in winter, there are four to five such
+systems in the Northern Hemisphere alone [@Zhu1998], accounting for
 $80-90 \,\%$ of the total north-south integrated vapor transport
-`[@Guan2015; @Zhu1998]`.  Its dual hydrological role, both as a fresh
-water source for some water-stressed areas `[@Dettinger2011;
-@Dettinger2013; @Rutz2012]` and as a potential trigger for floods
-`[@Lavers2012; @Lavers2013; @Neiman2008; @Moore2012]`, has granted it
+[@Guan2015; @Zhu1998].  Its dual hydrological role, both as a fresh
+water source for some water-stressed areas [@Dettinger2011;
+@Dettinger2013; @Rutz2012] and as a potential trigger for floods
+[@Lavers2012; @Lavers2013; @Neiman2008; @Moore2012], has granted it
 increasing attentions among the research community.  And their
 long-term change in a warming climate also stands as pressing research
 question. However, an important prerequisite to answer such questions
@@ -53,11 +53,11 @@ consistency and objectivity compared with human
 identification.
 
 In many existing applications, a magnitude thresholding
-approach is used. For instance, `@Ralph2004`, `@Neiman2008`,
-`@Hagos2015` and `@Dettinger2011` identified ARs by first locating
+approach is used. For instance, @Ralph2004, @Neiman2008,
+@Hagos2015 and @Dettinger2011 identified ARs by first locating
 regions where the Integrated Water Vapor (IWV) is greater than $20\,
 mm$.  A $250 \, kg/m/s$ threshold on the Integrated Vapor Transport
-(IVT) was used by `@Rutz2014` and `@Rutz2015`.
+(IVT) was used by @Rutz2014 and @Rutz2015.
 However, an implicit assumption
 with this magnitude thresholding approach is that the atmospheric
 moisture level stays unchanged throughout the analysis period.  Such
@@ -67,7 +67,7 @@ the atmospheric moisture level is expected to increase.
 In this package we propose a suite of new detection/tracking
 to help solve the above difficulties.
 Through a
-systematic analysis using 7 years of Reanalysis data `[@Dee2011]`, we
+systematic analysis using 7 years of Reanalysis data [@Dee2011], we
 have found that the proposed detection algorithm has reduced sensitivity to
 parameters and data resolution.  Long-lived ARs spanning multiple
 days, having cross-continent or cross-basin tracks can be more reliably
@@ -95,7 +95,7 @@ The AR detection algorithm is inspired and modified from the image
 processing algorithm *Top-hap by Reconstruction (THR)*, which consists
 of subtracting from the original image a *greyscale reconstruction by
 dilation* image.  Some more details of the THR algorithm and its
-applications can be found in this work of `[@Vincent1993]`.
+applications can be found in this work of [@Vincent1993].
 
 In the context of AR detection, the greyscale image in question is the
 non-negative IVT distribution, denoted as $I$.  The greyscale
[…]
