### 2020-05-27 #2197 @Xunius
@kbarnhart Thanks. I think I must have accidentally reverted to a previous commit or something, I remembered myself fixing these citation issues before. Anyway, I just pushed a new commit fixing this, and tried the online preview service again. Now it compiles correctly.
