### 2020-05-27 c5e4443
capitalize joss paper title
