### 2020-05-28 #2197 @kbarnhart
@Xunius

Thank you for submitting this contribution to JOSS. As part of the review process, one of the JOSS editors (me) does an initial screening to ensure the submission is in scope and identify if there are any issues that should be addressed before the review process starts. I am very excited to see this contribution reviewed at JOSS, but based on examining the contents of the submission and discussing with the JOSS editorial board, there are a number of things that I think should be addressed before the submission proceeds to review. I've outlined these issues below in a checklist.

As always, please do not hesitate to ask questions or raise concerns. You may contact me here on this issue thread or at krbarnhart@usgs.gov

The typical procedure at JOSS at this point is to label the submission as "paused" while you address these issues. When you are done, you would notify me and I will resume the submission.

- [ ] _Installation instructions:_ The contribution is a set of scripts without formal packaging. Formal packaging is not required, but some instructions for how a user might use this package outside of the source directory is merited. In addition, I might consider making a conda-environment specification.
- [ ] _Example Usage:_ The set of notebooks you have created provide nice examples but they appear to rely on data from a Google Drive. I would recommend either providing the example data as part of the package or pointing to an existing reanalysis archive.
- [ ] _API documentation:_ While it looks like everything that is exposed in a function signature is documented, examining the code itself reveals that there are quite a few local and global variables that are hard coded into the scripts. These include paths to files and the contents of `PARAM_DICT`. I would recommend that these parameters exposed as user specified without the need to edit the source code (e.g., as function keyword arguments).
- [ ] _Tests:_ It appears that tests exist but there are no instructions for running them.
- [ ] _Atmospheric river algorithm context:_ Could you place your algorithm more clearly in the context of the [ARTMIP effort](http://www.cgd.ucar.edu/projects/artmip/) (Rutz et al., 2019; Shields et al., 2018) and the algorithms considered by that group.

_References_
Rutz, J. J., Shields, C. A., Lora, J. M., Payne, A. E., Guan, B., Ullrich, P., et al. ( 2019). The atmospheric river tracking method intercomparison project (ARTMIP): quantifying uncertainties in atmospheric river climatology. Journal of Geophysical Research: Atmospheres, 2019; 124: 13777– 13802. https://doi.org/10.1029/2019JD030936

Shields, C. A., Rutz, J. J., Leung, L.-Y., Ralph, F. M., Wehner, M., Kawzenuk, B., Lora, J. M., McClenny, E., Osborne, T., Payne, A. E., Ullrich, P., Gershunov, A., Goldenson, N., Guan, B., Qian, Y., Ramos, A. M., Sarangi, C., Sellars, S., Gorodetskaya, I., Kashinath, K., Kurlin, V., Mahoney, K., Muszynski, G., Pierce, R., Subramanian, A. C., Tome, R., Waliser, D., Walton, D., Wick, G., Wilson, A., Lavers, D., Prabhat, Collow, A., Krishnan, H., Magnusdottir, G., and Nguyen, P.: Atmospheric River Tracking Method Intercomparison Project (ARTMIP): project goals and experimental design, Geosci. Model Dev., 11, 2455–2474, https://doi.org/10.5194/gmd-11-2455-2018, 2018.
