### 2020-06-01 #2197 @Xunius
@kbarnhart Thanks for the comments.
I'm not quite used to packaging the code as nicely as a professional programmer, because 99% of the time the only user of the code I wrote is myself, so there is no surprise there is a lot of rough edges.

I got a few questions:
1. I can't promise making a conda-environment specification will work, I'll have to try. My experience with the `cdat` package is that it can be annoying to install (and I'm blaming other science programmers for not packaging the code well).
2. I'm still not sure the best way to provide the sample data. The computation requires at least a few hundred MB of data, I doubt the user would like to clone a repo with this big of binary data in it. On the other hand, many reanalysis data centers require a sign-up to download, and the user has to select the data domain in the web interface to download. So a link to the data archive alone probably is not sufficient. I would also need to give some instructions on the selection of the required domain?
3. on the "instructions for running them" part, is `python -m unittest test/test_something.py` the correct type of instruction? Again I'm not used to write tests.
