### 2020-06-01 #2197 @kbarnhart
Thanks for those questions @Xunius I'll do my best to provide some answers. I would also recommend looking at some of repositories of [other papers](https://joss.theoj.org/) published in JOSS to see some existing examples of providing installation and testing instructions, as well as approaches to documentation and packaging code. I often find looking at these examples to be helpful.

- Just to be clear, I'm not recommending that you need to distribute your code through conda-forge or pypi, just to create a `environment.yml` specification so that the required environment is described by that file and a user might install it with something like the following ([conda environment documentation](https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#creating-an-environment-from-an-environment-yml-file)):

```bash
conda env create -f environment.yml
```

- The reason I recommend this is that one of the things that happens during the JOSS review process is that reviewers will test installation instructions. While it is not a requirement, using Continuous Integration platforms like Appveyor and Travis is one way to check that your installation instructions and tests work on multiple platforms and multiple python versions.

- Regarding sample data: I agree that the sample data should not be present in this repository. You make a good point regarding many re-analysis data centers requiring a log in. I will discuss with the JOSS editorial board and see if we can make a recommendation for how to best handle this.

- Regarding instructions for testing: I would recommend a sentence that describes what the existing tests do and a statement about how a user might run the all the tests. There are a number of ways you could do this (with unittest only, or with additional testing packages such as nosetest or pytest). The important thing is that you indicate to a user how to assess whether they have installed the package correctly and the what the tests passing means.

As always, please reach out with questions and for clarification.
