### 2020-06-02 2f8590a
add var.id=varid in readVar()
