### 2020-06-06 30ded87
refactor code WIP
