+++ scripts/detect_ARs.py
@@ -1,0 +1,263 @@
+"""Locate atmospheric rivers (ARs) from integrated water vapor transports
+(IVTs) using Top-hat by reconstruction (THR) algorithm.
+
+# Input data
+
+1. uflux, vflux:
+
+    6-hourly vertically integrated moisture flux, in kg/m/s.
+    Data should be formatted into 4D (time, singleton_z, latitude, longitue),
+    or 3D (time, latitude, longitude).
+
+2. IVT:
+
+    6-hourly THR reconstruction and anomalies of IVT, in kg/m/s.
+    File names:
+
+        ivt_m1-60_6_<year>_cln-minimal-rec-ano-kernel-t<T>-s<S>.nc
+
+    Where:
+        <year>: year in yyyy.
+        <T>: kernel size in time dimension.
+        <S>: kernel size in space (horizontal) dimension.
+
+    E.g.:
+
+        ivt_m1-60_6_2000_cln-minimal-rec-ano-kernel-t12-s8.nc
+
+    These are generated in compute_thr_singlefile.py or compute_thr_multifile.py
+
+    Data should be formatted into 4D (time, singleton_z, latitude, longitue),
+    or 3D (time, latitude, longitude).
+
+NOTE: data should have proper time, latitude and longitude axes!
+
+# Domain
+
+Take only northern hemisphere, shift longitude to 80 E.
+
+Author: guangzhi XU (xugzhi1987@gmail.com; guangzhi.xu@outlook.com)
+Update time: 2020-04-01 12:15:12.
+"""
+
+from __future__ import print_function
+
+
+
+#######################################################################
+#                               Globals                               #
+#######################################################################
+
+#--------------------Time range--------------------
+YEAR=2007
+TIME_START='%d-01-01 00:00:00' %YEAR
+TIME_END='%d-01-05 18:00:00' %YEAR
+
+#-----------u-qflux----------------------
+SOURCEDIR1='/home/guangzhi/datasets/erai_qflux/'
+UQ_FILE_NAME='uflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
+UQ_VAR='uflux'
+
+#-----------v-qflux----------------------
+SOURCEDIR2='/home/guangzhi/datasets/erai_qflux'
+VQ_FILE_NAME='vflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
+VQ_VAR='vflux'
+
+#-----------------ivt reconstruction and anomalies-----------------
+SOURCEDIR3='/home/guangzhi/datasets/erai/ERAI_AR_THR/'
+IVT_FILE_NAME='ivt_m1-60_6_%d_cln-cea-proj-THR-kernel-t16-s6.nc' %YEAR
+
+
+#------------------Output folder------------------
+OUTPUTDIR='/home/guangzhi/datasets/erai/ERAI_AR_THR/%d/' %YEAR
+LABEL_FILE_OUT_NAME='ar_s_6_%d_label-angle-flux.nc' %YEAR
+RECORD_FILE_OUT_NAME='ar_records_%d.csv' %YEAR
+
+
+
+PLOT=True          # create maps of found ARs or not
+
+LAT1=0; LAT2=90      # degree, latitude domain
+# NOTE: this has to match the domain seletion in compute_thr_singlefile.py
+
+RESO=0.75             # degree, (approximate) horizontal resolution of input data.
+SHIFT_LON=80          # degree, shift left bound to longitude. Should match
+                      # that used in compute_thr_singlefile.py
+
+PARAM_DICT={
+    # kg/m/s, define AR candidates as regions >= than this anomalous ivt.
+    'thres_low' : 1,
[…]
