### 2020-06-07 0e442b6
more reformat
rename AR_tracker folder to ipart

mv test_data.py, compute_ivt.py to scripts/ folder

rename detect_ARs2.py to detect_ARs_generator_version.py

doc changes in detect_ARs.py, detect_ARs_generator_version.py,
trace_ARs.py, AR_detector.py.
