+++ ipart/AR_detector.py
@@ -26,7 +26,7 @@ NX_VERSION=nx.__version__[0]
 
 
 
-def plotGraph(graph,ax=None,show=True):
+def plotGraph(graph, ax=None, show=True):
     '''Helper func to plot the graph of an AR coordinates
     '''
 
@@ -43,7 +43,7 @@ def plotGraph(graph,ax=None,show=True):
     return
 
 
-def areaFilt(mask,area,min_area=None,max_area=None):
+def areaFilt(mask, area, min_area=None, max_area=None):
     '''Filter AR binary masks by region areas
 
     Args:
@@ -77,7 +77,7 @@ def areaFilt(mask,area,min_area=None,max_area=None):
     return result
 
 
-def spherical2Cart(lat,lon):
+def spherical2Cart(lat, lon):
     clat=(90-lat)*np.pi/180.
     lon=lon*np.pi/180.
     x=np.cos(lon)*np.sin(clat)
@@ -87,7 +87,7 @@ def spherical2Cart(lat,lon):
     return np.array([x,y,z])
 
 
-def cart2Spherical(x,y,z, shift_lon):
+def cart2Spherical(x, y, z, shift_lon):
 
     r=np.sqrt(x**2+y**2+z**2)
     clat=np.arccos(z/r)/np.pi*180
@@ -100,7 +100,7 @@ def cart2Spherical(x,y,z, shift_lon):
     return np.array([lat,lon,np.ones(lat.shape)])
 
 
-def computeTheta(p1,p2):
+def computeTheta(p1, p2):
     r'''Tangent line to the arc \|p1-p2\|
 
     Args:
@@ -115,7 +115,7 @@ def computeTheta(p1,p2):
     return theta
 
 
-def wind2Cart(u,v,lats,lons):
+def wind2Cart(u, v, lats, lons):
     '''Convert u,v winds to Cartesian, consistent with spherical2Cart.
     '''
 
@@ -131,7 +131,7 @@ def wind2Cart(u,v,lats,lons):
     return vs
 
 
-def cart2Wind(vs,lats,lons):
+def cart2Wind(vs, lats, lons):
     '''Convert winds in Cartesian to u,v, inverse to wind2Cart.
     '''
 
@@ -1167,8 +1167,6 @@ def crossSectionFlux(mask, quslab, qvslab, axis_rdp):
     return angles,anglesmean,crossflux,seg_thetas
 
 
-
-
 def findARAxis(quslab, qvslab, armask_list, costhetas, sinthetas, param_dict,
         verbose=True):
     '''Find AR axis
@@ -1213,6 +1211,58 @@ def findARAxis(quslab, qvslab, armask_list, costhetas, sinthetas, param_dict,
     return axes, axismask
 
 
+def prepareMeta(lats, lons, times, ntime, nlat, nlon,
+        ref_time='days since 1900-01-01', verbose=True):
+    '''Prepare metadata for AR detection function calls
+
+    Args:
+        lats (ndarray): 1D, latitude coordinates, the length needs to equal
+            <nlat>.
+        lons (ndarray): 1D, longitude coordinates, the length needs to equal
+            <nlon>.
+        times (list or array): time stamps of the input data as a list of strings,
+            e.g. ['2007-01-01 06:00:00', '2007-01-01 12:00', ...].
+            Needs to have the a length of <ntime>.
+
+    Keyword Args:
+        ref_time (str): reference time point to create time axis.
+
+    Returns:
+        timeax (cdms2.axis.TransientAxis): a time axis obj created from strings
+            in <times>.
+        areamap (ndarray): grid cell areas in km^2, with shape (<nlat> x <nlon>).
[…]
