### 2020-06-08 66a0001
update docs, minor print changes in notebooks
