+++ docs/river_tracker2.rst
@@ -1,7 +1,0 @@
-Documentation page for river_tracker2.py
-========================================
-
-.. automodule:: river_tracker2
-  :members:
-
-
