+++ docs/thr.rst
@@ -1,0 +1,6 @@
+Documentation page for thr.py
+=============================
+
+.. automodule:: thr
+  :members:
+
