+++ docs/funcs.rst
@@ -2,5 +2,5 @@ Documentation page for utils/funcs.py
 =====================================
 
 .. automodule:: utils.funcs
-  :members: dLatitude, dLongitude, readVar
+  :members: get3DEllipse, dLatitude, dLatitude2, dLongitude, dLongitude2, readVar, getTimeAxis
 
