+++ docs/index.rst
@@ -1,9 +1,9 @@
-.. AR tracker documentation master file, created by
+.. IPART documentation master file, created by
    sphinx-quickstart on Sun Mar 29 17:13:48 2020.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-AR tracker's documentation
+IPART's documentation
 ==========================
 
 .. contents:: Table of Contents
@@ -13,10 +13,11 @@ AR tracker's documentation
 Introduction
 ############
 
-**AR tracker** is a Python package for automated Atmospheric River (AR)
-detection, axis finding and AR tracking from gridded Integrated Vapor Transport (IVT) data, for
-instance `Reanalysis
-datasets <https://www.esrl.noaa.gov/psd/data/gridded/reanalysis/>`_.
+**IPART** (Image-Processing based Atmospheric River Tracking) is a Python
+package for automated Atmospheric River (AR) detection, axis finding and AR
+tracking from gridded Integrated Vapor Transport (IVT) data, for instance
+`Reanalysis datasets
+<https://www.esrl.noaa.gov/psd/data/gridded/reanalysis/>`_.
 
 An overview of what ARs are can be found in this review paper: `Atmospheric
 rivers: a mini-review <https://doi.org/10.3389/feart.2014.00002>`_.
@@ -42,21 +43,25 @@ Below is an example output figure:
 Installation
 ############
 
-This package at the moment does not support installation via ``pip`` or ``conda``.
-Users are advised to build a working Python virtual environment with required `Dependencies`_, and obtain
-the code of this pacakge from the `github page <https://github.com/ihesp/AR_tracker>`_. 
+The package can be installed via ``conda``.
+Users are advised to build a working Python virtual environment with the
+required `Dependencies`_, and install the package by ``conda install -c guangzhi ipart``.
+Alternatively, one can obtain the code of this pacakge from the
+`github page <https://github.com/ihesp/IPART>`_, and add the ``ipart``
+module to the ``PYTHONPATH`` environment variable.
 
 
+Detailed Installation Guide
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-
-Recommend building the Python environment using `Anaconda <https://www.anaconda.com/distribution/>`_.
+Recommend building the Python environment using
+`Anaconda <https://www.anaconda.com/distribution/>`_.
 
 After Anaconda installation, create a working environment:
 ::
 
     conda create -n ar python=3.7
     conda activate ar
-
 
 This creates a Python environment named "ar" with Python version 3.7.
 
@@ -66,8 +71,9 @@ Then install the above listed dependencies, e.g.
     conda install numpy
 
 
-This installs the current latest version of ``numpy``. Most likely the latest versions will work, in case of compatibility issues, consider forcing a given version of a package, e.g.
-::
+This installs the current latest version of ``numpy``. Most likely the latest
+versions will work, in case of compatibility issues, consider forcing a given
+version of a package, e.g.  ::
 
     conda install matplotlib=2.2.3
 
@@ -82,13 +88,11 @@ For installation of ``CDAT``, checkout the `installation guides <https://github.
 
 If nothing prints out, the installation is successful. In case of errors, also consider their `partial installation instructions <https://github.com/CDAT/cdat/wiki/Additional-Installation-Configurations>`_. Only the ``cdms2`` and ``cdutil`` modules are needed, the ``vcs`` module is not required.
 
-Lastly, get a copy of this repository, either by doing a ``git clone``:
+Lastly, install the ``ipart`` package
 ::
 
[…]
