+++ docs/river_tracker1_funcs.rst
@@ -1,8 +1,0 @@
-Documentation page for river_tracker1_funcs.py
-==============================================
-
-.. automodule:: river_tracker1_funcs
-  :members:
-
-
-
