+++ docs/compute_thr_singlefile.rst
@@ -1,5 +1,0 @@
-Documentation page for compute_thr_singlefile.py
-================================================
-
-.. automodule:: compute_thr_singlefile
-  :members:
