+++ docs/AR_detector.rst
@@ -1,0 +1,6 @@
+Documentation page for AR_detector.py
+=====================================
+
+.. automodule:: AR_detector
+  :members:
+

[…]
