### 2020-06-08 11c2ad6
squeeze inputs in findARsGen()
AR_detector.py:

    in findARsGen(), squeeze ivt, ivtrec, ivtano, qu, qv inputs to 3D.

detect_ARs.py, detect_ARs_generator_version.py:

    remove the `timeax=None` test code.
