### 2020-06-08 62bc8dd
rename test folder to tests. fix some tests. README change
