+++ joss/paper.bib
@@ -633,3 +633,26 @@ url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2019JD031218},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2019JD031218},
 year = {2019}
 }
+
+@article{Rutz2019b,
+    author = {Rutz, Jonathan J. and Shields, Christine A. and Lora, Juan M. and Payne, Ashley E. and Guan, Bin and Ullrich, Paul and O’Brien, Travis and Leung, L. Ruby and Ralph, F. Martin and Wehner, Michael and Brands, Swen and Collow, Allison and Goldenson, Naomi and Gorodetskaya, Irina and Griffith, Helen and Kashinath, Karthik and Kawzenuk, Brian and Krishnan, Harinarayan and Kurlin, Vitaliy and Lavers, David and Magnusdottir, Gudrun and Mahoney, Kelly and McClenny, Elizabeth and Muszynski, Grzegorz and Nguyen, Phu Dinh and Prabhat, Mr. and Qian, Yun and Ramos, Alexandre M. and Sarangi, Chandan and Sellars, Scott and Shulgina, T. and Tome, Ricardo and Waliser, Duane and Walton, Daniel and Wick, Gary and Wilson, Anna M. and Viale, Maximiliano},
+    doi = {10.1029/2019JD030936},
+    pages = {13777--13802},
+    journal = {Journal of Geophysical Research: Atmospheres},
+    title = {The Atmospheric River Tracking Method Intercomparison Project (ARTMIP): Quantifying Uncertainties in Atmospheric River Climatology},
+    volume = {124},
+    year = {2019}
+}
+
+@article{Shields2018,
+    author = {Shields, Christine A. and Rutz, Jonathan J. and Leung, Lai-Yung and Ralph, F. Martin and Wehner, Michael and Kawzenuk, Brian and Lora, Juan M. and McClenny, Elizabeth and Osborne, Tashiana and Payne, Ashley E. and Ullrich, Paul and Gershunov, Alexander and Goldenson, Naomi and Guan, Bin and Qian, Yun and Ramos, Alexandre M. and Sarangi, Chandan and Sellars, Scott and Gorodetskaya, Irina and Kashinath, Karthik and Kurlin, Vitaliy and Mahoney, Kelly and Muszynski, Grzegorz and Pierce, Roger and Subramanian, Aneesh C. and Tome, Ricardo and Waliser, Duane and Walton, Daniel and Wick, Gary and Wilson, Anna and Lavers, David and Collow, Allison and Krishnan, Harinarayan and Magnusdottir, Gudrun and Nguyen, Phu},
+    doi = {10.5194/gmd-11-2455-2018},
+    issn = {1991-9603},
+    issue = {6},
+    pages = {2455-2474},
+    journal = {Geoscientific Model Development},
+    publisher = {Copernicus GmbH},
+    title = {Atmospheric River Tracking Method Intercomparison Project (ARTMIP): project goals and experimental design},
+    volume = {11},
+    year = {2018}
+}
