### 2020-06-08 #2197 @kbarnhart
@Xunius I wanted to follow up with some recommendations regarding the sample data based on discussing this with the rest of the JOSS editorial team.

We would recommend that you place the sample data in a public archive (e.g., zenodo, figshare) and then provide instructions for users to obtain the data (e.g., use wget/curl in the notebook).

I would also recommend indicating a link to the data archive (e.g., on whichever reanalysis server you are using) and a brief statement of what the data domain you have selected is.

Please let me know if I can provide any clarification on this recommendation.
