### 2020-06-08 93990b8
fix setup.py, tested local conda build and install
