### 2020-06-09 ad6ed5d
make sure lat increasing in findARsGen()
