### 2020-06-09 0f723de
minor fix in tests/test_river_tracker1.py. ready to merge
