### 2020-06-09 e863a51
fix import in ipart/__init__.py, minor fix in 3_detect_ARs.ipynb
