+++ ipart/__init__.py
@@ -1,4 +1,4 @@
-import AR_detector
-import AR_tracer
-import thr
-import utils
+from . import AR_detector
+from . import AR_tracer
+from . import thr
+from . import utils
