### 2020-06-09 8685c0f
try to fix sphinx build, NX_VERSION global causing issue?
