+++ docs/conf.py
@@ -12,6 +12,7 @@
 #
 import os
 import sys
+sys.path.insert(0, os.path.abspath('..'))
 sys.path.insert(0, os.path.abspath('../ipart'))
 sys.path.insert(0, os.path.abspath('../scripts'))
 
