+++ ipart/AR_detector.py
@@ -22,7 +22,6 @@ from ipart.utils import rdp
 from ipart.utils import funcs
 from ipart.utils import peak_prominence2d as pp2d
 
-NETWORKX_VERSION=nx.__version__[0]
 
 
 
@@ -344,7 +343,8 @@ def getARAxis(g, quslab, qvslab, mask):
         '''Sum edge distances along a path'''
         s=0
         for ii in range(len(path)-1):
-            if NETWORKX_VERSION=='2':
+            nxversion=nx.__version__[0]
+            if nxversion=='2':
                 sii=g[path[ii]][path[ii+1]][attr]
             else:
                 sii=g.edge[path[ii]][path[ii+1]][attr]
