### 2020-06-09 060cd6b
pillow vulnerability fix according to github
