### 2020-06-09 #2197 @Xunius
@kbarnhart Thanks for the recommendation.

We have made some changes to the submission:

1. We are changing to the title to "IPART: A Python Package for Image-Processing based Atmospheric River Tracking". The github repo name is also changed to https://github.com/ihesp/IPART (I noticed that clicking the old URL at the top of this issue/thread points to the updated repo though).
2. The package is now install-able via conda: `conda install -c guangzhi ipart`. I've uploaded a py2 and a py3 version to anaconda cloud, and tested both on my own machine. The relevant installation instructions are also updated accordingly in the README.md and the readthedocs page.
3. Code has been re-structured so that the core functions are put in the `ipart` module, and a few example scripts are given in the `scripts` folder, mainly as templates for users to modify and build their own working scripts.
4. Notebooks have been updated using the new module design.
5. I've managed to squeeze the sample data to a much smaller size (~10 MB each, ~20 MB total), and I think that's acceptable size to put in the repo itself. So they are now saved in the `notebooks` folder. Metadata of these sample data are given in the relevant notebook.
6. The readthedocs pages are updated.
7. One requirement file is created for py2 version and another for py3 version.
8. Some minor changes to the tests, and in the README.md, added the instruction to run the tests: `python -m unittest discover -s tests`.

Please let me know whether these changes sufficiently address the previous issues.
