### 2020-06-11 #2197 @kbarnhart
@Xunius thank you for your update. Unfortunately, I don't think your changes sufficiently address the issues I pointed out in the checklist above. Based on the changes you've made I notice two additional concerns.

- [ ] I am concerned by your use of conda-forge to provide the package,  without using conda to mange the dependencies. Please correct me if I am wrong, but it appears that your conda binary doesn't manage the dependencies. JOSS intends to be flexible regarding the specifics of software installation in order to respect the differences between submissions and the needs of the developers, but use of a package manager _implies_ that the package manager is managing dependencies. I would recommend that you do either (a), or (a) and (b) of the following:

    a) Provide a conda [environment .yml](https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#creating-an-environment-from-an-environment-yml-file) file with instructions to use it to create the environment. After creating and activating the environment, a user would then be instructed to run something like `python setup.py` or `pip install -e .` to install the package from source. Thus installation instructions could be something similar to the following:

    ```bash
    git clone https://github.com/ihesp/IPART.git
    cd ipart
    conda env create -f environment.yml
    python setup.py install
    ```

    b) If you want to distribute a binary through a package manager (which is certainly encouraged but not is not required), please ensure that installation by this method manages the dependencies.

- [ ] It does not appear to me that you have described how your algorithm compares with those presented in the two ARTMIP papers. I see the addition of the two references, but no discussion of algorithms.

I want to follow up regarding the comment I made above about hard coded paths and the `PARAM_DICT` argument. I made this comment because one of the [submission requirements for JOSS](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) is that a submission must be _"feature-complete (no half-baked solutions) and designed for maintainable extension (not one-off modifications)"_. The presence of hard coded paths and a large number of input parameters that are not documented as part of the API indicated to me that the package has not yet reached a state where it is designed for maintainable extension. I think that your change to distinguish between the `ipart` module and the set of example scripts is a good first step.

One of the goals of JOSS is to improve software through the review process, so I provided the above feedback in support of improving this submission so it can meet this criteria and move forward into review.

I look forward to seeing this submission meet these minimum requirements and move forward to review. However, at present, it is not yet there.

Please let me know if you have any questions or clarifications. I will do my best to answer them.
