### 2020-06-11 #2197 @kbarnhart
@openjournals/joss-eics could you please update the repository address and title of this submission? I don't think I have privileges to do this.

Title: "IPART: A Python Package for Image-Processing based Atmospheric River Tracking"
Repo: https://github.com/ihesp/IPART
