### 2020-06-12 a07fa5c
fix the flawed PDF issue in plot.py
