### 2020-06-12 5440257
make the PARAM_DICT params kwards in findARs(), findARsGen()
