+++ environment_py3.yml
@@ -1,0 +1,246 @@
+name: ipart3
+channels:
+  - guangzhi
+  - cdat/label/v81
+  - conda-forge
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=main
+  - alabaster=0.7.12=py_0
+  - astroid=2.4.2=py37hc8dfbb8_0
+  - attrs=19.3.0=py_0
+  - autopep8=1.5.3=pyh9f0ad1d_0
+  - babel=2.8.0=py_0
+  - backcall=0.1.0=py_0
+  - basemap=1.2.0=py37h705c2d8_0
+  - blas=1.1=openblas
+  - bleach=3.1.5=pyh9f0ad1d_0
+  - bokeh=2.0.1=py37hc8dfbb8_0
+  - brotlipy=0.7.0=py37h8f50634_1000
+  - bzip2=1.0.8=h516909a_2
+  - ca-certificates=2020.4.5.2=hecda079_0
+  - cdat=8.1=1
+  - cdat_info=8.1.1=py_2
+  - cdp=1.4.2=py_1
+  - cdtime=3.1.2=py37h7eb8c7e_6
+  - cdutil=8.1=py_1
+  - certifi=2020.4.5.2=py37hc8dfbb8_0
+  - cffi=1.14.0=py37hd463f26_0
+  - chardet=3.0.4=py37hc8dfbb8_1006
+  - cibots=0.2=py_0
+  - click=7.1.2=pyh9f0ad1d_0
+  - cloudpickle=1.4.1=py_0
+  - cryptography=2.9.2=py37hb09aad4_0
+  - curl=7.69.1=h33f0ec9_0
+  - cycler=0.10.0=py37_0
+  - cytoolz=0.10.1=py37h7b6447c_0
+  - dask=2.17.2=py_0
+  - dask-core=2.17.2=py_0
+  - dbus=1.13.14=hb2f20db_0
+  - decorator=4.4.2=py_0
+  - defusedxml=0.6.0=py_0
+  - distarray=2.12.2=py_1
+  - distributed=2.18.0=py37hc8dfbb8_0
+  - docutils=0.16=py37hc8dfbb8_1
+  - dv3d=8.1=py_0
+  - entrypoints=0.3=py37hc8dfbb8_1001
+  - eofs=1.4.0=py_0
+  - esmf=7.1.0=h9eb252b_1005
+  - esmpy=7.1.0=py37h5ca1d4c_3
+  - expat=2.2.6=he6710b0_0
+  - ffmpeg=4.1.3=h167e202_0
+  - flake8=3.8.2=pyh9f0ad1d_0
+  - fontconfig=2.13.0=h9420a91_0
+  - freetype=2.9.1=h8a8886c_1
+  - fsspec=0.7.4=py_0
+  - future=0.18.2=py37hc8dfbb8_1
+  - g2clib=1.6.0=hf3f1b0b_9
+  - genutil=8.1.1=py37h3010b51_0
+  - geos=3.6.2=heeff764_2
+  - ghostscript=9.22=hf484d3e_1001
+  - glib=2.63.1=h5a9c865_0
+  - gmp=6.2.0=he1b5a44_2
+  - gnutls=3.6.13=h79a8f9a_0
+  - gst-plugins-base=1.14.0=hbbd80ab_1
+  - gstreamer=1.14.0=hb31296c_0
+  - hdf4=4.2.13=hf30be14_1003
+  - hdf5=1.10.4=nompi_h3c11f04_1106
+  - heapdict=1.0.1=py_0
+  - icu=58.2=he6710b0_3
+  - idna=2.9=py_1
+  - imageio=2.8.0=py_0
[…]
