### 2020-06-12 28e6af7
change installation in README and docs
