+++ joss/paper.md
@@ -72,6 +72,12 @@ high-latitude landfalls. As the research on ARs matures, new AR
 detection/tracking methods are been developed, and the inter-comparisons of
 various AR detection/tracking are carried out by, for instance, the Atmospheric
 River Tracking Method Intercomparison Project (ARTMIP) [@Rutz2019b;@Shields2018].
+We have performed some systematic comparisons with two magnitude thresholding based AR detection methods included in ARTMIP, and the proposed method displays
+better correspondence between North Hemisphere AR tracks and
+storm tracks, better identification of the strong mid-latitude
+AR-related moisture transports, and longer AR track durations. The detailed
+analysis has been accepted by the journal of Atmosphere and is now in
+production.
 
 The *IPART* package includes a collection of Python functions/classes designed
 for an analysis workflow covering the detection of ARs, the simplification of
