+++ README.md
@@ -17,14 +17,51 @@
 
 Recommend building the Python environment using [Anaconda](https://www.anaconda.com/distribution/).
 
+### Option 1, create conda env using environment file
+
+After Anaconda installation, git clone this repository:
+
+```
+git clone https://github.com/ihesp/IPART
+```
+
+Then build a new conda environment using the environment file provided. For
+Python 2.7:
+
+```
+cd IPART
+conda env create -f environment_py2.yml
+```
+
+This creates a new environment named `ipartpy2`. Activate the environment using
+
+```
+conda activate ipartpy2
+```
+
+After that, you can check the list of packages installed by
+
+```
+conda list
+```
+
+Similarly for Python 3.7, use
+
+```
+conda env create -f environment_py3.yml
+```
+
+
+### Option 2, manually install all packages
+
 After Anaconda installation, create a working environment:
 
 ```
-conda create -n ar python=3.7
-conda activate ar
+conda create -n ipartpy3 python=3.7
+conda activate ipartpy3
 ```
 
-This creates a Python environment named "ar" with Python version 3.7.
+This creates a Python environment named "ipartpy3" with Python version 3.7.
 
 Then install the above listed dependencies, e.g.
 
