+++ environment_py2.yml
@@ -1,4 +1,4 @@
-name: ipart2
+name: ipartpy2
 channels:
   - guangzhi
   - defaults
