+++ environment_py3.yml
@@ -1,4 +1,4 @@
-name: ipart3
+name: ipartpy3
 channels:
   - guangzhi
   - cdat/label/v81
