### 2020-06-12 e14878d
add reso to return docs
