### 2020-06-12 #2197 @Xunius
@kbarnhart
Based on your comments I have made some further changes:
1. added 2 environment files and instructions for creating the conda envs with them.
2. things in `PARAM_DICT` are now function keyword arguments, function calls are changed to `func(..., **PARAM_DICT)`. The dict name being spelled using capitals does not make it "hard coded", and I wonder how much of a difference this makes.
3. I also doubt that JOSS is the suitable platform for discussing methodological inter-comparisons like what ARTMIP is doing. I've added 2 sentences describing the comparisons we have made. The work was submitted to a separate journal, and is still in production atm, so I can't give a reference just yet.
