### 2020-06-12 #2197 @kbarnhart
@Xunius thank you for making those changes.

1. I see that you have created new environment files, however, your changes have not addressed my concern regarding use of a package manager but not managing dependencies. Specifically, your environment.yml files still point to the conda-forge distribution of IPART that does not manage the dependencies.

    I will re-state what I said in the prior comment, that use of a package manager is encouraged, but not required, and if a package manager is used, it should handle the dependencies.

    I also note that your conda environment files appear to be an un-edited export from `conda env export` such that every package installed in your environment is listed and your computer's prefix is listed. I recommend that you use `conda env export --from-history` as described [here](https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#exporting-an-environment-file-across-platforms) and edit the .yml file so that one might expect it would work on another person's computer (for example, make sure to remove the prefix section of the yaml). The environment.yml only needs to list the direct dependencies of IPART.

2. Thank you for revising the function signature so that the parameters that control AR detection are documented. You have now sufficiently addressed this issue.

    A related clarification: When I mentioned "hard coded" I was referring to the hard coded paths to files on your file system in your example scripts (for example, [here](https://github.com/ihesp/IPART/blob/master/scripts/detect_ARs.py#L99), not to the `PARAM_DICT`.

    The issue with the `PARAM_DICT` (that you have now addressed) is that the parameters that control AR detection were not described or documented in the API documentation. Describing these parameters in the API documentation _does make a difference_.

3. Regarding comparison to ARTMIP, perhaps I have not been clear in my recommendation. I agree that JOSS is not a suitable platform for methodological inter-comparisons. When I wrote _"Could you place your algorithm more clearly in the context of the ARTMIP effort (Rutz et al., 2019; Shields et al., 2018) and the algorithms considered by that group."_ I was trying to encourage you to describe how your method fits into the classification of AR algorithms used by the ARTMIP project. For example, Ruiz et al. (2019) Table 1 and Figure 5 classify the AR algorithms on a number of different bases (e.g., type, geometry requirements, threshold requirements,  absolute/relative, global/regional). As best as I can tell, your contribution is a relative threshold method, by their classification. But this is just one aspect of the classification. Two new sentences that place your algorithm in the context of their classification scheme would be more than sufficient.

Please let me know if you have questions or clarifications about my comments. I look forward to hearing from you that these issues have been addressed so that this submission can move forward into review.
