### 2020-06-13 67a4a02
rebuild environment files. fix tests
