+++ environment_py2.yml
@@ -1,151 +1,14 @@
 name: ipartpy2
 channels:
-  - guangzhi
   - defaults
+  - conda-forge
 dependencies:
-  - _libgcc_mutex=0.1=main
-  - attrs=19.3.0=py_0
-  - backports=1.0=py_2
-  - backports.functools_lru_cache=1.6.1=py_0
-  - backports.shutil_get_terminal_size=1.0.0=py27_2
-  - backports_abc=0.5=py27_0
-  - basemap=1.2.0=py27h705c2d8_0
-  - blas=1.0=mkl
-  - bleach=3.1.4=py_0
-  - bzip2=1.0.8=h7b6447c_0
-  - ca-certificates=2020.1.1=0
-  - cdat-lite=6.0rc2=py27_1
-  - certifi=2019.11.28=py27_0
-  - cftime=1.0.4.2=py27hdd07704_0
-  - cloudpickle=1.2.2=py_0
-  - configparser=4.0.2=py27_0
-  - contextlib2=0.6.0.post1=py_0
-  - curl=7.69.1=hbc83047_0
-  - cycler=0.10.0=py27_0
-  - cytoolz=0.10.1=py27h7b6447c_0
-  - dask-core=1.2.2=py_0
-  - dbus=1.13.14=hb2f20db_0
-  - decorator=4.4.2=py_0
-  - defusedxml=0.6.0=py_0
-  - entrypoints=0.3=py27_0
-  - enum34=1.1.6=py27_1
-  - expat=2.2.6=he6710b0_0
-  - fontconfig=2.13.0=h9420a91_0
-  - freetype=2.9.1=h8a8886c_1
-  - functools32=3.2.3.2=py27_1
-  - futures=3.3.0=py27_0
-  - geos=3.6.2=heeff764_2
-  - glib=2.63.1=h3eb4bd4_1
-  - gmp=6.1.2=h6c8ec71_1
-  - gst-plugins-base=1.14.0=hbbd80ab_1
-  - gstreamer=1.14.0=hb31296c_0
-  - hdf4=4.2.13=h3ca952b_2
-  - hdf5=1.10.4=hb1b8bf9_0
-  - icu=58.2=he6710b0_3
-  - imageio=2.6.1=py27_0
-  - importlib_metadata=1.3.0=py27_0
-  - intel-openmp=2020.1=217
-  - ipaddress=1.0.23=py_0
-  - ipart=2.0=py27_0
-  - ipykernel=4.10.0=py27_0
-  - ipython=5.8.0=py27_0
-  - ipython_genutils=0.2.0=py27_0
-  - ipywidgets=7.5.1=py_0
-  - jinja2=2.11.2=py_0
-  - jpeg=9b=h024ee3a_2
-  - jsonschema=3.2.0=py27_0
-  - jupyter=1.0.0=py27_7
-  - jupyter_client=5.3.4=py27_0
-  - jupyter_console=5.2.0=py27_1
-  - jupyter_core=4.6.1=py27_0
-  - kiwisolver=1.1.0=py27he6710b0_0
-  - krb5=1.17.1=h173b8e3_0
-  - libcurl=7.69.1=h20c2e04_0
-  - libedit=3.1.20181209=hc058e9b_0
-  - libffi=3.3=he6710b0_1
-  - libgcc-ng=9.1.0=hdf63c60_0
-  - libgfortran-ng=7.3.0=hdf63c60_0
-  - libnetcdf=4.6.1=h11d0813_2
-  - libpng=1.6.37=hbc83047_0
-  - libsodium=1.0.16=h1bed415_0
-  - libssh2=1.9.0=h1ba5d50_1
[…]
