### 2020-06-13 #2197 @Xunius
@kbarnhart please checkout the new changes:
1. the environment files are re-created using `--from-history` flag, and without the `prefix`. It only includes direct dependencies.
2. Minor edits in `paper.md`.
