### 2020-06-15 d8d41c6
remove two requirements.txt files that list every package in the environment
(only direct dependencies should be listed)
