### 2020-06-15 a3d91d1
Add a note to state platform support
