### 2020-06-15 43ed419
remove sections that refer to conda channel installation
removed because the conda package does not manage dependencies.
