### 2020-06-15 #2197 @kbarnhart
@Xunius Thank you for updating me to these changes. I think the added text related to ARTMIP substantially improves the paper.

Regarding the installation, your Option 1 does not actually install IPART. This linked pull request makes some recommended changes to your setup.py file and readme.md. ihesp/IPART/pull/1

Regarding your installation Options 2 and 3 (which I have removed in the PR), I note that you have ignored the statement I made twice in prior comments that it is not appropriate to use conda to install a package without also using it to also manage that packages dependencies. This issue must be addressed before this submission can move forward into review. One way to address it would be to delete the files on anaconda.org and to merge my PR. You, of course, may address it however you see fit.

You might wonder why I have continued to comment on the installation instructions. The reason is as follows: It is very important that papers submitted to JOSS have installation instructions that reviewers can be expected to follow. One of the first items in the JOSS review checklist is:

**Installation:** Does installation proceed as outlined in the documentation?

It is for this reason that, as a topical editor, I carefully check the installation instructions during the pre-review process.

Please let me know if you have questions or clarifications about my comments. I look forward to hearing from you that this issue has been addressed so that this submission can move forward into review.
