### 2020-06-16 #2197 @Xunius
@kbarnhart Thanks. The PR has been merged.
The conda package came with dependencies in my previous fix, except that for py3, the cdms2 and cdutil dependencies don't work out of the box ([one related issue I'm having](https://github.com/CDAT/cdat/issues/2252#issuecomment-644428720)), that's why I added manual install instructions if that one doesn't work. For py2 it works as expected, installing `ipart` from conda pulls in the dependencies.

As for the previous option 1, are you sure that didn't install IPART? It works on my machine without doing `python setup` or `pip install`. Anything to do with conda version? I'm using 4.8.2.
