### 2020-06-16 #2197 @kbarnhart
@Xunius you can import a python module _when you are in the code's directory_ without running setup.py develop. That is why option 1 appeared to have installed ipart. If you are in another directory, however, it won't import.

Regarding conda forge: Just in case I'm not being clear enough, what I mean by _manage all dependencies_ is that one might expect the following to install ipart and all of its dependencies:
```
$ conda create --name ipart
$ conda activate ipart
$ conda install -c guangzhi ipart
```
such that the following would work
```
$ python
>>> import ipart
```

I find the following:

**Ubuntu 18.04**
```
$ conda create -n ipart
$ conda activate ipart
$ conda install -c guangzhi ipart

Collecting package metadata (current_repodata.json): done
Solving environment: failed with initial frozen solve. Retrying with flexible solve.
Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
Solving environment: failed with initial frozen solve. Retrying with flexible solve.
Solving environment: \
Found conflicts! Looking for incompatible packages.
This can take several minutes.  Press CTRL-C to abort.
failed

UnsatisfiableError: The following specifications were found to be incompatible with each other:

Output in format: Requested package -> Available versions
```

**OSX**
```
$ conda create -n ipart
$ conda activate ipart
$ conda install -c guangzhi ipart
$ python
Python 3.7.6 | packaged by conda-forge | (default, Jun  1 2020, 18:33:30)
[Clang 9.0.1 ] on darwin
>>> ipart
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
NameError: name 'ipart' is not defined
>>> import ipart
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/barnhark/opt/anaconda3/envs/ipart/lib/python3.7/site-packages/ipart-2.0-py3.7.egg/ipart/__init__.py", line 1, in <module>
  File "/Users/barnhark/opt/anaconda3/envs/ipart/lib/python3.7/site-packages/ipart-2.0-py3.7.egg/ipart/AR_detector.py", line 8, in <module>
ModuleNotFoundError: No module named 'numpy'
>>> import numpy
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ModuleNotFoundError: No module named 'numpy'
>>>
```
The conda installation fails on ubuntu (where I was able to successfully install using the revised Option 1), appears to succeed on Mac, but when one tries to import ipart, I find that the dependencies have not been managed (as the dependency numpy has not been installed).

I recognize that you indicated in that "the conda package came with dependencies in my previous fix". Would you have expected my above installation to work?

While we wrap up addressing this installation issue, could you also provide the GitHub handles of  reviewers you think would be appropriate for this submission? Please mention their GitHub handles without the @ (you would refer to me as kbarnhart) so they don't get unnecessary emails.
