### 2020-06-16 #2197 @Xunius
@kbarnhart I'm able to create and install `ipart` all in one line, but only in py2. Doing this in separate steps seems much harder. For py3 it the conda install doesn't work. I'll have to look into that later.

For suggested reviews, perhaps:
- ajdawson
- doutriaux1
- jwblin
