### 2020-06-19 #2197 @muszyna25
Hi Katy,

Thank you for your invitation to review the submission for JOSS. However, I am not able to review it, but I can recommend somebody that may be interested: @ burlen @ taobrienlbl

Thanks,
Grzegorz
