### 2020-06-24 fe3aa19
try make zonally cyclic
AR_detector.py:

    in areaFilt(), add kwarg zonal_cyclic=False to pass into
    cyclicLabel().

    add checkCyclic() to check if a binary mask is zonally cyclic.

    in getARData(), call checkCyclic(maskii) before getting contour,
    shift longitude if so.

    in getARData(), when examining qvmeanii, multiply by -1 if SH.

    add breakCurveAtEdge(), break line curves across cyclic zonal edges.

    in plotAR(), call breakCurveAtEdge() and plot each segment
    separately.

    in findARAxis(), roll data if zonally cyclic to place data at
    center.

    add cycl
[…]
