+++ scripts/detect_ARs_generator_version.py
@@ -91,27 +91,34 @@ from __future__ import print_function
 #######################################################################
 
 #--------------------Time range--------------------
-YEAR=2007
+YEAR=1980
 TIME_START='%d-01-01 00:00:00' %YEAR
-TIME_END='%d-01-05 18:00:00' %YEAR
+TIME_END='%d-01-20 18:00:00' %YEAR
 
 #-----------u-qflux----------------------
-SOURCEDIR1='/home/guangzhi/datasets/erai_qflux/'
-UQ_FILE_NAME='uflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
-UQ_VAR='uflux'
+#SOURCEDIR1='/home/guangzhi/datasets/erai_qflux/'
+#UQ_FILE_NAME='uflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
+UQ_FILE_NAME='/home/guangzhi/datasets/artmip_merra_added_time/ivt_s_3_1980_merra2-NH.xml'
+#UQ_VAR='uflux'
+UQ_VAR='uIVT'
 
 #-----------v-qflux----------------------
-SOURCEDIR2='/home/guangzhi/datasets/erai_qflux'
-VQ_FILE_NAME='vflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
-VQ_VAR='vflux'
+#SOURCEDIR2='/home/guangzhi/datasets/erai_qflux'
+#VQ_FILE_NAME='vflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
+#VQ_VAR='vflux'
+VQ_FILE_NAME=UQ_FILE_NAME
+VQ_VAR='vIVT'
 
 #-----------------ivt reconstruction and anomalies-----------------
-SOURCEDIR3='/home/guangzhi/datasets/erai/ERAI_AR_THR/'
-IVT_FILE_NAME='ivt_m1-60_6_%d_cln-cea-proj-THR-kernel-t16-s6.nc' %YEAR
+#SOURCEDIR3='/home/guangzhi/datasets/erai/ERAI_AR_THR/'
+#IVT_FILE_NAME='ivt_m1-60_6_%d_cln-cea-proj-THR-kernel-t16-s6.nc' %YEAR
+SOURCEDIR3='/home/guangzhi/datasets/artmip_merra_added_time/THR'
+IVT_FILE_NAME='ivt_s_3_%d_merra2-NH-THR-kernel-t32-s9.nc' %YEAR
 
 
 #------------------Output folder------------------
-OUTPUTDIR='/home/guangzhi/datasets/erai/ERAI_AR_THR/%d/' %YEAR
+#OUTPUTDIR='/home/guangzhi/datasets/erai/ERAI_AR_THR/%d/' %YEAR
+OUTPUTDIR='/home/guangzhi/datasets/artmip_merra_added_time/'
 LABEL_FILE_OUT_NAME='ar_s_6_%d_label-angle-flux.nc' %YEAR
 RECORD_FILE_OUT_NAME='ar_records_%d.csv' %YEAR
 
@@ -119,8 +126,8 @@ RECORD_FILE_OUT_NAME='ar_records_%d.csv' %YEAR
 
 PLOT=True          # create maps of found ARs or not
 
-LAT1=0; LAT2=90      # degree, latitude domain
-SHIFT_LON=80          # degree, shift left bound to longitude.
+LAT1=-90; LAT2=0      # degree, latitude domain
+SHIFT_LON=0          # degree, shift left bound to longitude.
 
 PARAM_DICT={
     # kg/m/s, define AR candidates as regions >= than this anomalous ivt.
@@ -134,9 +141,9 @@ PARAM_DICT={
[…]
