### 2020-06-26 30897a5
more zonal cyclic adjusts. add docs
AR_detector.py:

    in getARData(), fix the AR contour computation. Shift the mask but
    not the lonax, after getting the mask, shift lon coordinates based
    on which side the central longitude. This removes the weird jumpping
    curves across the map edges.

    add determineThresLow(), an empirical function to determine the
    thres_low for anoslab.

    in _findARs(), fix the latitudinal centroid filtering to work for
    both NH and SH. Roll the mask before peak partition. Roll the mask
    before filling mask holes.

    add docstrings to new fun
[…]
