+++ ipart/AR_detector.py
@@ -148,7 +148,17 @@ def cart2Wind(vs, lats, lons):
 
     return u,v
 
+
 def checkCyclic(mask):
+    '''Check binary mask is zonally cyclic or not
+
+    Args:
+        mask (ndarray): 2d binary array, with 1s denoting existance of
+            a feature.
+    Returns:
+        result (bool): True if feature in <mask> is zonally cyclic, False
+            otherwise.
+    '''
 
     left=mask[:,0]
     right=mask[:,-1]
@@ -156,6 +166,7 @@ def checkCyclic(mask):
         return True
     else:
         return False
+
 
 def maskToGraph(mask, quslab, qvslab, costhetas, sinthetas, edge_eps,
         connectivity=2):
@@ -744,13 +755,17 @@ def getARData(slab, quslab, qvslab, anoslab, quano, qvano, areas,
 
         # mask contour
         if checkCyclic(maskii):
+            # if mask is zonally cyclic, shift to center
             maskii_roll=np.roll(maskii, maskii.shape[1]//2, axis=1)
-            lonax_roll=np.roll(lonax[:], len(lonax)//2)
-            # NOTE: holes exist in contours across zonal edges
-            contii=funcs.getBinContour(maskii_roll,lonax_roll,latax)
+            contii=funcs.getBinContour(maskii_roll,lonax,latax)
+            # have to shift the longitude coordinates again
             contii=contii.vertices
-            #newx=(contii[:,0] - maskii.shape[1]//2)%maskii.shape[1]
-            #contii[:,0]=newx
+            xx=contii[:,0]
+            xx2=np.zeros(xx.shape)
+            mid_lon=lonax[len(lonax)//2]
+            xx2=np.where(xx<=mid_lon, xx+180, xx2)
+            xx2=np.where(xx>mid_lon, xx-180, xx2)
+            contii[:,0]=xx2
         else:
             contii=funcs.getBinContour(maskii,lonax,latax)
             contii=contii.vertices
@@ -815,6 +830,7 @@ def getARData(slab, quslab, qvslab, anoslab, quano, qvano, areas,
             is_relaxedii=True
         if lenii<min_length:
             is_relaxedii=True
+        # NOTE that in SH, qvmean<0 is poleward
         if np.sign(centroidy)*qvmeanii<=0:
             is_relaxedii=True
 
@@ -942,22 +958,39 @@ def save2DF(result_dict):
 
     return result_df
 
-def breakCurveAtEdge(xs, ys, left_bound, right_bound):
 
-    idx=[]
-    new_xs=[]
+def breakCurveAtEdge(xs, ys, left_bound, right_bound):
+    '''Segment curve coordinates at the left, right edges
+
+    Args:
+        xs (ndarray): 1d array of x- coordinates.
+        ys (ndarray): 1d array of y- coordinates.
+        left_bound (float): left most bound of the map domain.
+        right_bound (float): right most bound of the map domain.
+    Returns:
+        new_xs (list): list of 1d arrays, each being a segment of the
+            original input <xs>.
+        new_ys (list): list of 1d arrays, each being a segment of the
+            original input <ys>.
+
+    This function segment a curve's coordinates into a number of segments
+    so that when plotted in a basemap plot, a zonally cyclic curve won't
+    be plotted as jumping straight lines linking the left and right bounds.
+    '''
+
+    idx=[]  # break point indices
+    new_xs=[] # result list for x coordinates segments
     new_ys=[]
     for ii, xii in enumerate(xs[:-1]):
         xii2=xs[ii+1]
-        #if (xii-left_bound)*(xii2-left_bound)<=0 or\
-                #(xii-right_bound)*(xii2-right_bound)<=0:
-        dx=abs(xii2-xii)
[…]
