### 2020-06-26 9e16ced
make track plots zonally cyclic. readme change
AR_detector.py:

    move breakCurveAtEdge() to utils.funcs.py.
    some docstring change.

AR_tracer.py:

    in plotAR(), use latax, lonax range to define map domain.
    break curve segments using breakCurveAtEdge() before plotting.

funcs.py:

    move breakCurveAtEdge() to utils.funcs.py.

trace_ARs.py:

    test with MERRA data

test_river_tracker1.py:

    add zonal_cyclic to param_dict.

README.md:

    add changelog
