### 2020-06-27 #2197 @sadielbartholomew
Hi Katy, thank you for offering a review. Yes, I am happy & able to review this submission & do so at least within 6 weeks. I have read the COI Policy & do not have any COI for this submission.

One question to confirm, though, as I have not reviewed for JOSS before: I assume it does not preclude me from being a reviewer overall if I am due to soon be a co-author on a submission to JOSS (completely unrelated to this submission)? It's okay for reviewers also to submit to JOSS generally? Thanks.
