### 2020-06-28 f67b836
AR_detector.py:
    rename partPeaks() to partPeaksOld(), make a new partPeaks() using
    area to filter peaks, and watershed to part peaks.

    in getARData(), compute end points distance and if too close, skip.
    If L/W < 2, skip. Don't filter using max_isoq_hard anymore.
    isoperimetric quotient estimate doesn't seem to be quite accurate.

    add kwarg sill=0.8 to determineThresLow().

    in _findARs(), call the new partPeaks(), don't filter using
    max_isoq_hard.

funcs.py:

    add getGridsInContour().

peak_prominence2d.py:

    in getProminence(), comment out prints.

compute_
[…]
