### 2020-06-29 #2197 @kbarnhart
@sadielbartholomew thank you for being willing to review this submission! I can confirm that soon-to-be submitting to JOSS does not preclude you from being a reviewer.

I'll add you to the submission as a reviewer. I hope to start the official review thread in the next few days.

Thanks again for participating in the JOSS review process.
