### 2020-06-29 #2197 @kbarnhart
👋 @andreas-h and @rabernat a quick ping to consider reviewing this submission to JOSS.
