### 2020-06-29 #2407 @whedon
[REVIEW]: IPART: A Python Package for Image-Processing based Atmospheric River Tracking
**Submitting author:** @Xunius (<a href="http://orcid.org/0000-0001-5877-5608">Guangzhi XU</a>)
**Repository:** <a href="https://github.com/ihesp/IPART" target ="_blank">https://github.com/ihesp/IPART</a>
**Version:** v3.0.8
