### 2020-06-29 #2407 @kbarnhart
@arfon / @openjournals/joss-eics could you help me ensure the repository is linked correctly to this review issue. Issues may have cropped up because the authors changed the title of the submission and the repository address during the pre-review stage (#2197).

I can edit the title/comments on this thread... @whedon could find the paper in the review issue (e.g., [here](https://github.com/openjournals/joss-reviews/issues/2197#issuecomment-644209397))

repo: https://github.com/ihesp/IPART
title: "IPART: A Python Package for Image-Processing based Atmospheric River Tracking"

thanks!
