### 2020-06-30 280f4cf
AR_detector.py:
    in partPeaks(), change way of separating watershed results.

    in getARData(), remove max_isoq_hard and add min_LW. Remove end
    points distance filtering.

    in _findARs(), remove max_isoq_hard. Use partPeaks() instead of
    partPeaksOld().

    in findARs() and findARsGen(), remove max_isoq_hard, add min_LW.

compute_thr_singlefile.py:

    test run.

detect_ARs.py, detect_ARs_generator_version.py, test_river_tracker1.py:

    remove max_isoq_hard add min_LW.
