### 2020-06-30 80954b7
move max_iso and iso_quotient
