### 2020-06-30 9cc5853
remove isoperimeter from docs
