### 2020-06-30 a49cf32
Revert "remove isoperimeter from docs"
This reverts commit 9cc58539600db9ff9a92d8ea6653813abd47f995.
