### 2020-07-17 a732a83
use fill_radius//2 when breaking merges
