### 2020-07-21 cc0d078
try to remove cdat dependency
new files with a suffix of "2" are modified version with no cdat
dependencies. Use instead netcdf4.

thr.py:

    use netcdf4 to replace cdat.

NOTICE_TO_ME.txt: functions need changes
