+++ ipart/utils/funcs2.py
@@ -1,0 +1,1670 @@
+'''Utility functions
+
+Author: guangzhi XU (xugzhi1987@gmail.com)
+Update time: 2020-06-05 23:23:22.
+'''
+from __future__ import print_function
+import copy
+from datetime import datetime, timedelta
+import numpy as np
+#import pandas as pd
+from netCDF4 import Dataset, date2num, num2date
+
+
+class DataError(Exception):
+    def __init__(self, string=None):
+        if string != None:
+            self.message = string
+    def __str__(self):
+        return self.message
+
+def isInteger(x):
+    """Check an input is integer
+
+    Args:
+        x (unknow type): input
+
+    Returns: True if <x> is integer type, False otherwise.
+    """
+
+    import sys
+    import numpy
+
+    if sys.version_info.major==2:
+        return isinstance(x, (int, long, numpy.integer))
+    else:
+        return isinstance(x, (int, numpy.integer))
+
+class Selector(object):
+    def __init__(self, v1, v2, axis=0):
+        #if v1>v2:
+            #v1, v2= v2, v1
+        self.v1=v1
+        self.v2=v2
+        self.axis=axis
+
+class NCVAR(object):
+    def __init__(self, data, id=None, axislist=None, attributes=None):
+
+        if id is None:
+            id='unnamed'
+
+        self.id=str(id)
+        self.data=data
+
+        if axislist is None:
+            axislist=[]
+            for ii in range(data.ndim):
+                axisii=createAxis(ii, np.arange(data.shape[ii]).astype('f'))
+                axislist.append(axisii)
+
+        if attributes is None:
+            attributes={'id': str(id),
+                    'name': str(id),
+                    'long_name': '',
+                    'standard_name': '',
+                    'title': '',
+                    'units': ''}
+
+        self._axislist=axislist
+        self._attributes=attributes
+        for kk, vv in attributes.items():
+            setattr(self, kk, vv)
+
+    def __str__(self):
+        return self.data.__str__()
+
+    def __repr__(self):
+        return self.data.__repr__()
+
+    def __getitem__(self, idx):
+        if isInteger(idx):
+            return self.sliceIndex(idx, idx+1)
+        elif isinstance(idx, (list, tuple)):
+            if all([isinstance(ii, slice) for ii in idx]):
+                result=self
+                for ii, sii in enumerate(idx):
+                    if sii.start is None and sii.stop is None:
+                        continue
+                    result=result.sliceIndex(sii.start, sii.stop, axis=ii)
+                return result
+
+    def __call__(self, selectors):
+        if selectors is None:
+            return self
+        if isinstance(selectors, Selector):
+            selectors=(selectors,)
+        if isinstance(selectors, (list, tuple)) and\
+                all([isinstance(ii, Selector) for ii in selectors]):
+
+            result=self
+            for ii, sii in enumerate(selectors):
+                if sii.v1 is None and sii.v2 is None:
+                    continue
+                result=result.sliceData(sii.v1, sii.v2, sii.axis)
+
+            return result
+        else:
+            raise Exception("Input needs to be a tuple of Selectors.")
+
+
+
+    @property
+    def axislist(self):
+        return copy.deepcopy(self._axislist)
+
+    @axislist.setter
+    def axislist(self, newlist):
+        self._axislist=newlist
+
+    @property
[…]

[…]
