### 2020-07-21 ef9396b
change tests to use netcdf4 version
