### 2020-07-22 105bfae
doc changes to reflect the cdat->netCDF4 transiation
