+++ ipart/utils/plot2.py
@@ -1,7 +1,7 @@
 '''Plotting Functions.
 
 Author: guangzhi XU (xugzhi1987@gmail.com)
-Update time: 2017-10-09 14:32:05.
+Update time: 2020-07-22 09:27:30.
 '''
 
 #--------Import modules--------------
@@ -128,9 +128,10 @@ def mkscale(n1,n2,nc=12,zero=1):
 def index2Letter(index,verbose=True):
     '''Translate an integer index to letter index
 
-    <index>: integer index for a subplot.
-
-    Return <letter>: corresponding letter index for <index>.
+    Args:
+        index (int): integer index for a subplot.
+    Returns:
+        letter (str): corresponding letter index for <index>.
 
     <index> to letter indexing is defined as following:
     ----------------------------------------------------
@@ -181,46 +182,46 @@ def index2Letter(index,verbose=True):
     else:
         return index_dict[index]
 
-
 #--------------Get a matplotlib blue-to-red colormap obj--------------
 def colormapBR(reverse=False,verbose=True):
     '''Get a matplotlib blue-to-red colormap obj
 
-    <reverse>: bool, if True, colormap goes from red-white-blue.
-
-    Return <cmap>: matplotlib colormap.
-    Update time: 2015-10-25 16:09:53.
+    Args:
+        reverse (bool): if True, colormap goes from red-white-blue.
+    Returns:
+        cmap (matplotlib colormap): a blue-white-red colormap.
     '''
 
     import matplotlib.pyplot as plt
-
     if reverse:
         return plt.cm.bwr_r
     else:
         return plt.cm.bwr
 
-
 #-------------Map color indices----------------------
 def mapColor(levels,cmap,split=0,verbose=True):
     '''Map a list of levels linearly onto a colormap.
 
-    <levels>: a list of level values;
-    <cmap>: matplotlib colormap obj;
-    <split>: int, control behavior of negative and positive values\
-            0: Do not split negatives and positives, map onto entire\
-               range of [0,1];
-            1: split only when vmin<0 and vmax>0, otherwise\
-               map onto entire range of [0,1];
-               If split is to be applied, negative values are mapped onto first half [0,0.5],
-               and postives onto second half (0.5,1].
-            2: force split, if vmin<0 and vmax>0, same as <split>==1;
-                If vmin<0 and vmax<0, map onto 1st half [0,0.5];
-                If vmin>0 and vmax>0, map onto 2nd half (0.5,1].
-
-    Return: <colors>: a list of matplotlib colors sampled from <cmap>,
-                      each corresponds to a level in <levels>.
+    Args:
+        levels (1darray): a array/list/tuple of level values.
+        cmap (matplotlib colormap): colormap obj.
+    Keyword Args:
+        split (int): controls the behavior of negative and positive values
+                0: Do not split negatives and positives, map onto entire
+                   range of [0,1];
+                1: split only when vmin<0 and vmax>0, otherwise
+                   map onto entire range of [0,1];
+                   If split is to be applied, negative values are mapped onto
+                   first half [0,0.5], and postives onto second half (0.5,1].
+                2: force split, if vmin<0 and vmax>0, same as <split>==1;
+                    If vmin<0 and vmax<0, map onto 1st half [0,0.5];
+                    If vmin>0 and vmax>0, map onto 2nd half (0.5,1].
+    Returns:
+        colors (list): a list of matplotlib colors sampled from <cmap>,
+                       each corresponds to a level in <levels>.
 
     Usage:
+        ```
         cmap=plt.cm.jet
         xx=numpy.arange(12)
         colors=plot.mapColor(range(10),cmap,split=0)
@@ -228,8 +229,7 @@ def mapColor(levels,cmap,split=0,verbose=True):
             yy=numpy.sin(xx)+ii
[…]
