### 2020-07-22 d522226
notebook changes
README.md:

    dependency change.

AR_detector.py, funcs.py, plot.py, detect_ARs.py,
detect_ARs_generator_version.py:

    remove the old version, rename the new files.

change notebooks to use new functions.

setup.py:

    remove cdat modules, add netcdf4 to install_requires.
