+++ setup.py
@@ -12,12 +12,11 @@ setup(name='ipart',
         author_email='xugzhi1987@gmail.com',
         url='https://github.com/ihesp/IPART',
         install_requires=[
-            "cdms2",
+            "netcdf4",
             "matplotlib",
             "scipy",
             "scikit-image",
             "pandas",
-            "cdutil",
             "basemap",
             "networkx",
         ],
