+++ ipart/utils/funcs2.py
@@ -1,1838 +1,0 @@
-'''Utility functions
-
-Author: guangzhi XU (xugzhi1987@gmail.com)
-Update time: 2020-07-22 09:27:36.
-'''
-from __future__ import print_function
-import sys, os
-import copy
-from datetime import datetime, timedelta
-import numpy as np
-import pandas as pd
-from netCDF4 import Dataset, date2num, num2date
-
-
-class DataError(Exception):
-    def __init__(self, string=None):
-        if string != None:
-            self.message = string
-    def __str__(self):
-        return self.message
-
-def isInteger(x):
-    """Check an input is integer
-
-    Args:
-        x (unknow type): input
-    Returns:
-        True if <x> is integer type, False otherwise.
-    """
-    if sys.version_info.major==2:
-        return isinstance(x, (int, long, np.integer))
-    else:
-        return isinstance(x, (int, np.integer))
-
-class Selector(object):
-    def __init__(self, v1, v2, axis=0):
-        '''Similar to slice to slice NCVAR variable'''
-        self.v1=v1
-        self.v2=v2
-        self.axis=axis
-
-
-class NCVAR(object):
-    def __init__(self, data, id=None, axislist=None, attributes=None):
-        '''An object to store netcdf data and metadata
-
-        Args:
-            data (ndarray): data array.
-        Keyword Args:
-            id (str or None): id of data. If None, use a default "unnamed".
-            axislist (tuple or list or None): a tuple/list of NCVAR objs
-                acting as axes of the data. The length should equal the rank
-                of <data>. If None, create one axis for each dimension using
-                integer indices.
-            attributes (dict or None): variable attribute dictionary. If None,
-                create a new dict filled with some basic attributes.
-
-        This is a rudimentary attempt to mimic the API of CDAT package in
-        handling netcdf data. The netCDF4 API, imo, goes against the philosiphy
-        of of netcdf format. The data and metadata, after read in using
-        netCDF4, are saved to different variables, rather than bandled
-        together. One has to maintain these metadata himself, and it gets
-        bloody annoying quickly when things get complicated.
-
-        This class will try to save data and metadata together, and provide
-        some utility methods for interacting via metadata.
-
-        Examples:
-            * to create a variable:
-                ```
-                var = NCVAR(data_array, 'sst', (times, lats, lons),
-                     {'name': 'sst,
-                      'long_name': 'sea surface temperature',
-                      'standard_name': 'sea surface temperature',
-                      'units': 'K'})
-                ```
-            * to get the np data array:
-                ```var.data```
-            * to query the shape:
-                ```
-                var.shape, var.ndim, len(var)
-                ```
-            * to get the time axis:
-                ```var.getTime()```
-            * to get the latitude axis:
-                ```var.getLatitude()```
-            * to slice by latitude:
-                ```slab = var.sliceData(10, 60, axis=1)```
-            * to get the 1st time slice:
-                ```slab = var[0]```
-            * to slice by indices:
-                ```slab = var.sliceIndex(1, 10, axis=2)```
-                Alternatively, can slice by indices using a Selector:
-            * to create a selector:
-                ```sel = Selector(1, 10, axis=2)```
-            * to slice using the selector:
-                ```slab = var(sel)```
-                This will slice the data as ```data[:,:,1:10]```, the 2nd
[…]
