+++ scripts/detect_ARs_generator_version.py
@@ -34,8 +34,8 @@ and have compatible shapes.
 
 The user also needs to provide time, latitude and longitude axes metadata.
 In this script, these data are read in from the netCDF files using the
-CDAT (aka UVCDAT, CDAT8) package. If you are using some other package, e.g.
-netcdf4, xarray, iris or something else, please adjust the relevant code
+netCDF4 package. If you are using some other package, e.g.
+CDAT, xarray, iris or something else, please adjust the relevant code
 accordingly.
 
 # Domain
@@ -79,11 +79,10 @@ netCDF file.
     The boundary of all detected ARs are also marked out.
 
 Author: guangzhi XU (xugzhi1987@gmail.com)
-Update time: 2020-06-07 21:53:50.
+Update time: 2020-07-22 10:13:31.
 """
 
 from __future__ import print_function
-
 
 
 #######################################################################
@@ -91,37 +90,40 @@ from __future__ import print_function
 #######################################################################
 
 #--------------------Time range--------------------
-YEAR=1980
+YEAR=2007
 TIME_START='%d-01-01 00:00:00' %YEAR
-TIME_END='%d-01-20 18:00:00' %YEAR
+TIME_END='%d-01-10 18:00:00' %YEAR
 
 #-----------u-qflux----------------------
 #SOURCEDIR1='/home/guangzhi/datasets/erai_qflux/'
 #UQ_FILE_NAME='uflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
 UQ_FILE_NAME='/home/guangzhi/datasets/artmip_merra_added_time/ivt_s_3_1980_merra2-SH.xml'
-UQ_FILE_NAME='/home/guangzhi/datasets/artmip_merra_global/ivt_s_3_1980_merra2-global.xml'
-#UQ_VAR='uflux'
+UQ_FILE_NAME='/home/guangzhi/datasets/erai_qflux/uflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
 UQ_VAR='uIVT'
+UQ_VAR='uflux'
 
 #-----------v-qflux----------------------
 #SOURCEDIR2='/home/guangzhi/datasets/erai_qflux'
 #VQ_FILE_NAME='vflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
 #VQ_VAR='vflux'
 VQ_FILE_NAME=UQ_FILE_NAME
-VQ_VAR='vIVT'
+VQ_FILE_NAME='/home/guangzhi/datasets/erai_qflux/vflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
+VQ_VAR='vflux'
 
 #-----------------ivt reconstruction and anomalies-----------------
 #SOURCEDIR3='/home/guangzhi/datasets/erai/ERAI_AR_THR/'
 #IVT_FILE_NAME='ivt_m1-60_6_%d_cln-cea-proj-THR-kernel-t16-s6.nc' %YEAR
 SOURCEDIR3='/home/guangzhi/datasets/artmip_merra_added_time/THR'
 SOURCEDIR3='/home/guangzhi/datasets/artmip_merra_global/THR'
+SOURCEDIR3='/home/guangzhi/datasets/quicksave2/THR'
 IVT_FILE_NAME='ivt_s_3_%d_merra2-SH-THR-kernel-t32-s9.nc' %YEAR
-IVT_FILE_NAME='ivt_s_3_%d_merra2-global-THR-kernel-t32-s9.nc' %YEAR
+IVT_FILE_NAME='ivt_m1-60_6_%d_cln-cea-proj-THR-kernel-t10-s6.nc' %YEAR
 
 
[…]
