### 2020-07-22 ae1dab3
fix timeax in detect_ARs.py
README.md:

    add netcdf4 version for py3.

AR_detector.py:

    doc string change. AR boundary plot line width.

detect_ARs.py, detect_ARs_generator_version.py, trace_ARs.py:

    clear up file paths.

detect_ARs.py:

    use python datetime instead of cftime.
