+++ scripts/detect_ARs_generator_version.py
@@ -95,36 +95,19 @@ TIME_START='%d-01-01 00:00:00' %YEAR
 TIME_END='%d-01-10 18:00:00' %YEAR
 
 #-----------u-qflux----------------------
-#SOURCEDIR1='/home/guangzhi/datasets/erai_qflux/'
-#UQ_FILE_NAME='uflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
-UQ_FILE_NAME='/home/guangzhi/datasets/artmip_merra_added_time/ivt_s_3_1980_merra2-SH.xml'
 UQ_FILE_NAME='/home/guangzhi/datasets/erai_qflux/uflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
-UQ_VAR='uIVT'
 UQ_VAR='uflux'
 
 #-----------v-qflux----------------------
-#SOURCEDIR2='/home/guangzhi/datasets/erai_qflux'
-#VQ_FILE_NAME='vflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
-#VQ_VAR='vflux'
-VQ_FILE_NAME=UQ_FILE_NAME
 VQ_FILE_NAME='/home/guangzhi/datasets/erai_qflux/vflux_m1-60_6_%d_cln-cea-proj.nc' %YEAR
 VQ_VAR='vflux'
 
 #-----------------ivt reconstruction and anomalies-----------------
-#SOURCEDIR3='/home/guangzhi/datasets/erai/ERAI_AR_THR/'
-#IVT_FILE_NAME='ivt_m1-60_6_%d_cln-cea-proj-THR-kernel-t16-s6.nc' %YEAR
-SOURCEDIR3='/home/guangzhi/datasets/artmip_merra_added_time/THR'
-SOURCEDIR3='/home/guangzhi/datasets/artmip_merra_global/THR'
-SOURCEDIR3='/home/guangzhi/datasets/quicksave2/THR'
-IVT_FILE_NAME='ivt_s_3_%d_merra2-SH-THR-kernel-t32-s9.nc' %YEAR
-IVT_FILE_NAME='ivt_m1-60_6_%d_cln-cea-proj-THR-kernel-t10-s6.nc' %YEAR
-
+IVT_FILE_NAME='/home/guangzhi/datasets/quicksave2/THR/ivt_m1-60_6_%d_cln-cea-proj-THR-kernel-t10-s6.nc' %YEAR
 
 #------------------Output folder------------------
-#OUTPUTDIR='/home/guangzhi/datasets/erai/ERAI_AR_THR/%d/' %YEAR
-OUTPUTDIR='/home/guangzhi/datasets/artmip_merra_global/new'
 OUTPUTDIR='/home/guangzhi/datasets/quicksave2/THR'
-LABEL_FILE_OUT_NAME='ar_s_6_%d_label-angle-flux.nc' %YEAR
+LABEL_FILE_OUT_NAME='ar_label-angle-flux_%d.nc' %YEAR
 RECORD_FILE_OUT_NAME='ar_records_%d.csv' %YEAR
 
 
@@ -205,11 +188,9 @@ if __name__=='__main__':
     qvNV=qvNV.shiftLon(SHIFT_LON)
 
     #-------------------Read in ivt and THR results-------------------
-    abpath_in=os.path.join(SOURCEDIR3, IVT_FILE_NAME)
-    print('\n### <detect_ARs2>: Read in file:\n',abpath_in)
-    ivtNV=funcs.readNC(abpath_in, 'ivt')
-    ivtrecNV=funcs.readNC(abpath_in, 'ivt_rec')
-    ivtanoNV=funcs.readNC(abpath_in, 'ivt_ano')
[…]
