### 2020-07-23 18b670c
change environment files
README.md:

    update dependency versions.

update environment_py2.yml and environment_py3.yml to replace cdat with
netcdf4.

thr.py:

    fix ivt.data -> ivt

funcs.py:

    add num2dateWrapper()

compute_thr_singlefile.py:

    del unused line.

detect_ARs.py:

    use num2dateWrapper() instead of netCDF4.num2date
