+++ NOTICE_TO_ME.txt
@@ -1,44 +1,0 @@
-The netcdf branch is created from dev branch, not from master.
-This is to transit from CDAT to netcdf4.
-When done, merge back to dev.
-
-Functions need change:
-
-* [d][y]AR_detector.applyCropIdx()
-* [d][y]AR_detector.insertCropSlab()
-* [d][y]AR_detector.partPeaks()
-* [d][y]AR_detector.getARData()
-* [d][y]AR_detector.crossSectionFlux()
-* [d][y]AR_detector.prepareMeta()
-* [d][y]AR_detector.findARs()
-* [d][y]AR_detector.findARsGen()
-
-* [d][y]thr.THR()
-* [d][y]thr.THRCyclicLongitude()
-* [d][y]thr.rotatingTHR()
-
-* [d][y]utils.funcs.attribute_obj2dict()
-* [d][y]utils.funcs.attribute_dict2obj()
-* [d][y]utils.funcs.addExtraAxis()
-* [d][y]utils.funcs.cat()
-* [d][y]utils.funcs.interpretAxis()
-* [d][y]utils.funcs.getMissingMask() ?
-* [d][d]utils.funcs.getAxis()
-* [d][y]utils.funcs.increasingLatitude()
-* [d][y]utils.funcs.dLatitude2()
-* [d][y]utils.funcs.dLongitude2()
-* [d]utils.funcs.readVar() -> replace with readNC()
-* [d][y]utils.funcs.getTimeAxis()
-
-* [d][y]utils.plot.get1stTime()
-* [d][y]utils.plot.Plot2D
-* [d][y]utils.plot.Plot2Basemap
-
-* [d][y]scripts/compute_ivt.py
-* [d][y]scripts/compute_thr_singlefile.py
-* [d][y]scripts/compute_thr_multifile.py
-* [d][y]scripts/detect_ARs.py
-* [d][y[scripts/detect_ARs_generator_version.py
-* [d][y]scripts/test_data.py
-
-* [y]tests/*
