+++ environment_py3.yml
@@ -1,15 +1,15 @@
 name: ipartpy3
 channels:
-  - defaults
   - conda-forge
+  - defaults
 dependencies:
-  - cdms2
-  - matplotlib
-  - scipy
-  - scikit-image
-  - pandas
-  - cdutil
-  - basemap
   - python=3.7
+  - netcdf4
+  - numpy
+  - scipy
+  - basemap
+  - matplotlib=2.2.3
   - networkx
+  - pandas
+  - scikit-image
 
