### 2020-07-23 6d39bca
add doc string to funcs.num2dateWrapper()
