### 2020-07-23 #2407 @kbarnhart
@Xunius answer probably depends on what sort of changes you are making. I would not recommend totally restructuring the package. But if you have more moderate changes (adding a feature, updating the documentation, refactoring something internal), I don't see a reason why those sorts of changes would cause a problem. If you make a change that substantially modifies the dependencies such that a reviewer would need to re-create their compute environment used to test out and review IPART, that would probably not be advisable.

Using an issue to describe why changes were made, and then a Pull Request that bring those changes into master will also make it clear what changes were made and why.
