### 2020-07-24 91aee7e
Fixes to installation instructions in docs and README
