### 2020-07-24 43cc9f8
documentation update in response to issue #3
issue: https://github.com/ihesp/IPART/issues/3
