+++ ipart/AR_detector.py
@@ -26,7 +26,13 @@ from ipart.utils import peak_prominence2d as pp2d
 
 
 def plotGraph(graph, ax=None, show=True):
-    '''Helper func to plot the graph of an AR coordinates
+    '''Helper func to plot the graph of an AR coordinates. For debugging.
+
+    Args:
+        graph (networkx.Graph): networkx Graph obj to visualize.
+    Keyword Args:
+        ax (matplotlib axis obj): axis to plot on. If None, create a new one.
+        show (bool): whether to show the plot or not.
     '''
 
     if ax is None:
@@ -53,7 +59,6 @@ def areaFilt(mask, area, min_area=None, max_area=None):
                                   in <mask>.
         max_area (float or None): if not None, maximum area to filter objects
                                   in <mask>.
-
     Returns:
         result (ndarray): 2D binary mask with objects area-filtered.
     '''
@@ -77,6 +82,14 @@ def areaFilt(mask, area, min_area=None, max_area=None):
 
 
 def spherical2Cart(lat, lon):
+    '''Convert spherical to Cartesian coordiantes
+
+    Args:
+        lat,lon (float or ndarray): latitude and longitude coordinates.
+    Returns:
+        result (ndarray): Nx3 array, the columns are the x-, y- and z-
+            coordinates.
+    '''
     clat=(90-lat)*np.pi/180.
     lon=lon*np.pi/180.
     x=np.cos(lon)*np.sin(clat)
@@ -87,7 +100,16 @@ def spherical2Cart(lat, lon):
 
 
 def cart2Spherical(x, y, z, shift_lon):
+    '''Convert Cartesian to spherical coordiantes
 
+    Args:
+        x,y,z (float): x-, y- and z- coordiantes.
+        shift_lon (float): longitude to shift so the longitude dimension
+            starts from this number.
+    Returns:
+        result (ndrray): 1x3 array, the columns are the lat-, lon- and r-
+            coordinates. r- coordinates are all 1s (unit sphere).
+    '''
     r=np.sqrt(x**2+y**2+z**2)
     clat=np.arccos(z/r)/np.pi*180
     lat=90.-clat
@@ -104,6 +126,9 @@ def computeTheta(p1, p2):
 
     Args:
         p1,p2 (float): (lat,lon) coordinates
+    Returns:
+        theta (float): unit vector tangent at point <p1> pointing at <p2>
+            on unit sphere.
     '''
     p1=spherical2Cart(p1[0],p1[1])
     p2=spherical2Cart(p2[0],p2[1])
@@ -116,6 +141,14 @@ def computeTheta(p1, p2):
 
 def wind2Cart(u, v, lats, lons):
     '''Convert u,v winds to Cartesian, consistent with spherical2Cart.
+
+    Args:
+        u,v (float or ndarray): u- and v- components of horizontal winds.
+        lats,lons (float or ndarray): latitude and longitude coordinates
+            corresponding to the wind vectors given by <u> and <v>.
+    Returns:
+        vs (float or ndarray): Cartesian representation of the horizontal
+            winds.
     '''
 
     latsr=lats*np.pi/180
@@ -131,7 +164,15 @@ def wind2Cart(u, v, lats, lons):
 
 
 def cart2Wind(vs, lats, lons):
-    '''Convert winds in Cartesian to u,v, inverse to wind2Cart.
+    '''Convert winds in Cartesian coordinates to u,v, inverse to wind2Cart.
+
+    Args:
+        vs (float or ndarray): Cartesian representation of the horizontal
+            winds.
+        lats,lons (float or ndarray): latitude and longitude coordinates
+            corresponding to the wind vectors given by <u> and <v>.
[…]
