### 2020-07-24 ihesp/IPART#3 @Xunius
@sadielbartholomew Thanks for the comments.

Please checkout the [commit](https://github.com/ihesp/IPART/commit/43cc9f864ff0f190743ee9f41ce1f0836948e8fb) addressing these issues.

Regarding the documentation for other functions in `utils.func.py` and `utils.plot.py`, I added doc strings for them in another branch (`netcdf`), but I was [not advised](https://github.com/openjournals/joss-reviews/issues/2407#issuecomment-663039250) to merge until review is finished, as I made a dependency change in that branch.
