### 2020-07-24 0abdb8a
a bit more doc string change
