### 2020-07-24 ihesp/IPART#3 @sadielbartholomew
Thanks @Xunius. I will get back to you later today about this, & to continue my reviewing, but from what I can see from a quick look that all sounds good & I can likely close the Issue later on after the new commits you have made.
