### 2020-07-27 ihesp/IPART#3 @kbarnhart
Linking this to the JOSS review thread: openjournals/joss-reviews/issues/2407
