### 2020-07-27 #2407 @kbarnhart
@sadielbartholomew and @rabernat I wanted to check in with a friendly reminder that we are asking reviewers to complete their reviews within 6 weeks (in about 2 weeks). Please let me know how I can assist throughout the JOSS review process.
