### 2020-07-28 5b1cd11
re-implement doc changes made in master branch
master branch commit: https://github.com/ihesp/IPART/commit/0abdb8a98fbce55863c00254fab7dd42faef18c1
and
https://github.com/ihesp/IPART/commit/43cc9f864ff0f190743ee9f41ce1f0836948e8fb
