+++ ipart/thr.py
@@ -37,6 +37,7 @@ def THR(ivtNV, kernel, oroNV=None, high_terrain=600, verbose=True):
             Only used if <oroNV> is not None.
 
             New in v2.0.
+        verbose (bool): print some messages or not.
 
     Returns:
         ivtNV (NCVAR): 3D or 4D array, input <ivt>.
@@ -120,7 +121,9 @@ def rotatingTHR(filelist, varin, kernel, outputdir, oroNV=None,
                     make sure files in list have correct chronological order.
                     Note that time axis in data files should be the 1st axis.
         varin (str): variable id in files.
-        selector (Selector): selector obj to select subset of data.
+        kernel (list or tuple): list/tuple of integers specifying the shape of
+            the kernel/structuring element used in the gray erosion process.
+        selector (utils.funcs.Selector): selector obj to select subset of data.
         outputdir (str): path to folder to save outputs.
 
     Keyword Args:
@@ -140,6 +143,7 @@ def rotatingTHR(filelist, varin, kernel, outputdir, oroNV=None,
             Only used if <oroNV> is not None.
 
             New in v2.0.
+        verbose (bool): print some messages or not.
 
     Designed to perform temporal filtering on data that are too large to fit
     into memory, e.g. high-resolution data across multiple decades.
