+++ docs/plot.rst
@@ -1,6 +1,7 @@
 Documentation page for utils/plot.py
 ====================================
 
+Only some functions from this module are documented. The other undocumented functions are intended as private functions, not to be exposed to the user.
 
 .. automodule:: utils.plot
   :members: plot2
