+++ docs/funcs.rst
@@ -1,6 +1,8 @@
 Documentation page for utils/funcs.py
 =====================================
 
+Only some functions from this module are documented. The other undocumented functions are intended as private functions, not to be exposed to the user.
+
 .. automodule:: utils.funcs
   :members: get3DEllipse, dLatitude, dLatitude2, dLongitude, dLongitude2, readVar, getTimeAxis
 

[…]
