### 2020-07-28 #2407 @sadielbartholomew
Noted, thanks @kbarnhart. As this stage I do not feel I need assistance, I have simply not had much time in recent weeks to move on & finish my review. I plan, and will strive, to have my review completed by the end of this week i.e. the end of July.
