### 2020-07-29 f8bfa44
sphinx doc change to reflect the CDAT->netcdf4 dep change
