+++ docs/Detect-ARs.rst
@@ -12,11 +12,12 @@ An AR occurrence at a given time point is defined using these following rules:
 
 1. A connected region in the IVT anomaly field (:math:`I - \delta(I)`,
    computed in the section ":ref:`compute_thr`") where its values is greater than 0.
-2. The centroid (weighted by IVT values of the grid cells) of the region is north of :math:`20 ^{\circ} N`,
-   and south of :math:`80 ^{\circ}`, i.e. we are only interested in mid-latitude systems.
+2. The centroid (weighted by IVT values of the grid cells) of the region is north of :math:`20 ^{\circ} N` (or south of :math:`20 ^{\circ} S` for the Southern Hemisphere),
+   and south of :math:`80 ^{\circ} N` (or north of :math:`80 ^{\circ} S` for the Southern Hemisphere), i.e. we are only interested in mid-latitude systems.
 3. The region's area has to be within :math:`50 - 1800 \times 10^4 km^2`.
-4. After the computation of this AR candidates axis (see :ref:`compute_axis`) and the effective width (defined as area/length ratio), the length has to be :math:`\ge\, 1500 km`, and length/width ratio has to be :math:`\ge \,2`.
+4. After the computation of this AR candidates axis (see :ref:`compute_axis`) and the effective width (defined as area/length ratio), the length has to be :math:`\ge\, 1500 km`, and length/width ratio has to be :math:`\ge \,2` if length is below :math:`2000\, km`.
 
+.. note:: API is provided to control all these parameters.
 
 .. _detect_params:
 
@@ -48,7 +49,7 @@ Additional inputs:
             # km^2, drop AR candidates larger than this area.
             'max_area': 1800*1e4,
 
-            # float, min length/width ratio.
+            # float, min length/width ratio, applied only when length<min_length.
             'min_LW': 2,
 
             # degree, exclude systems whose centroids are lower than this latitude.
@@ -74,7 +75,7 @@ Additional inputs:
             'single_dome': False,
 
             # max prominence/height ratio of a local peak. Only used when SINGLE_DOME=True
-            'max_ph_ratio': 0.4,
+            'max_ph_ratio': 0.6,
 
             # minimal proportion of flux component in a direction to total flux to
             # allow edge building in that direction
@@ -91,25 +92,31 @@ The following snippet shows the detection function calls:
 ::
 
         from ipart.AR_detector import findARs
-        time_idx, labels, angles, crossfluxes, result_df = findARs(ivt, ivtrec,
-                    ivtano, qu, qv, latax, lonax, times=timeax **PARAM_DICT)
+        time_idx, labelsNV, anglesNV, crossfluxesNV, result_df = findARs(ivtNV.data,
+            ivtrecNV.data, ivtanoNV.data, quNV.data, qvNV.data, latax, lonax,
+            times=timeax, **PARAM_DICT)
 
 where these input arguments are:
 
-* ``ivt`` is the IVT data, with dimensions of ``(time, level, latitude, longitude)`` or ``(time, latitude, longitude)``.
-* ``ivtrec`` is :math:`\delta(I)`, and ``ivtano`` is :math:`I-\delta(I)`, see :ref:`compute_thr` for more details.
-* ``qu``: is :math:`F_u`, and ``qv`` is :math:`F_v`.
-* ``latax``: is an 1d array storing the latitude coordinates of ``ivt`` and others.
-* ``lonax``: is an 1d array storing the longitude coordinates of ``ivt`` and others.
+* ``ivtNV`` is a ``ipart.utils.NCVAR`` data object, which is a rudimentary wrapper
+  object designed to achieve a tighter bound between data values and metadata.
[…]
