+++ docs/Data-preparation.rst
@@ -73,59 +73,105 @@ The expected outputs would look like this:
 ::
 
         ### <test_data>: Read in file:
-         /home/guangzhi/datasets/erai/erai_qflux/uflux_m1-60_6_1984_cln.nc
+         /home/guangzhi/scripts/IPART/notebooks/uflux_s_6_1984_Jan.nc
         ##################################################
         Variable time axis:
         ##################################################
-        [1984-1-1 0:0:0.0, 1984-1-1 6:0:0.0, 1984-1-1 12:0:0.0, 1984-1-1
-        18:0:0.0, 1984-1-2 0:0:0.0, 1984-1-2 6:0:0.0, 1984-1-2 12:0:0.0,
-        ...
-        1984-12-30 12:0:0.0, 1984-12-30 18:0:0.0, 1984-12-31 0:0:0.0,
-        1984-12-31 6:0:0.0, 1984-12-31 12:0:0.0, 1984-12-31 18:0:0.0]
+        ### Description of slab ###
+          id: time
+          shape: (124,)
+          filename: None
+          missing_value: None
+          comments: None
+          grid_name: None
+          grid_type: None
+          long_name: time
+          units: hours since 1900-01-01 00:00:00.0
+          standard_name: None
+          Order: []
+
+        ### End of description ###
+
+        None
+        Time axis values:
+        [datetime.datetime(1984, 1, 1, 0, 0) datetime.datetime(1984, 1, 1, 6, 0)
+         datetime.datetime(1984, 1, 1, 12, 0) datetime.datetime(1984, 1, 1, 18, 0)
+         datetime.datetime(1984, 1, 2, 0, 0) datetime.datetime(1984, 1, 2, 6, 0)
+         datetime.datetime(1984, 1, 2, 12, 0) datetime.datetime(1984, 1, 2, 18, 0)
+         datetime.datetime(1984, 1, 29, 0, 0) datetime.datetime(1984, 1, 29, 6, 0)
+         ...
+         datetime.datetime(1984, 1, 31, 12, 0)
+         datetime.datetime(1984, 1, 31, 18, 0)]
 
         ##################################################
         Variable latitude axis:
         ##################################################
-           id: latitude
-           Designated a latitude axis.
-           units:  degrees_north
-           Length: 241
-           First:  -90.0
-           Last:   90.0
-           Other axis attributes:
-              long_name: latitude
-              axis: Y
-           Python id:  0x7f7b68896250
+        ### Description of slab ###
+          id: latitude
+          shape: (94,)
+          filename: None
+          missing_value: None
+          comments: None
+          grid_name: None
+          grid_type: None
+          long_name: latitude
+          units: degrees_north
+          standard_name: None
+          Order: []
 
+        ### End of description ###
+
+        None
+        Latitude axis values:
+        [10.   10.75 11.5  12.25 13.   13.75 14.5  15.25 16.   16.75 17.5  18.25
+         19.   19.75 20.5  21.25 22.   22.75 23.5  24.25 25.   25.75 26.5  27.25
+         28.   28.75 29.5  30.25 31.   31.75 32.5  33.25 34.   34.75 35.5  36.25
+         37.   37.75 38.5  39.25 40.   40.75 41.5  42.25 43.   43.75 44.5  45.25
+         46.   46.75 47.5  48.25 49.   49.75 50.5  51.25 52.   52.75 53.5  54.25
+         55.   55.75 56.5  57.25 58.   58.75 59.5  60.25 61.   61.75 62.5  63.25
+         64.   64.75 65.5  66.25 67.   67.75 68.5  69.25 70.   70.75 71.5  72.25
+         73.   73.75 74.5  75.25 76.   76.75 77.5  78.25 79.   79.75]
 
         ##################################################
         Variable longitude axis:
[…]
