### 2020-07-29 ihesp/IPART#5 @rabernat
Installation documentation and issues
This issue is in reference to my JOSS review of your paper (https://github.com/openjournals/joss-reviews/issues/2407).

I am trying to understand the recommended installation procedure for this package.

On one hand, you seem to provide a pip package with dependencies:
https://github.com/ihesp/IPART/blob/6bbf452b63b024011c8c67e442998d5d165200ae/setup.py#L14-L22

On the other hand, your [installation instructions](https://ipart.readthedocs.io/en/latest/index.html#installation) don't mention pip and instead recommend manually creating a new conda environment from scratch. The instructions are mostly about how to create conda environments. This does not adhere to best practices for distributing python code. For example, what if you want to use this package in an existing environment? Do you _have_ to create a new environment?

A better solution would be to provide working pip and / or conda-forge packages which automatically check for and, if needed, install the required dependencies on installation. This would allow the user to simply run `pip install ipart` or `conda install -c conda-forge ipart`. **Could you provide a pip or conda-forge package for this software?**

A related issue is whether the user is _required_ to clone the master repository in order to use the software. I notice that there are various scripts and notebooks that are part of the repo but not the package itself. Are these required or optional. Again, this is not a standard practice for scientific python software. **I recommend that users only be required to interact with the git repository if they wish to contribute to IPART development.** Otherwise a pip / conda install should be sufficient to use the software.

Making these changes would also help resolve #2.
