### 2020-07-29 ihesp/IPART#5 @rabernat
To support these suggestions, here are some references about how to properly package python software:

- https://ioos.github.io/ioos-python-package-skeleton/
- https://github.com/neurohackademy/nh2020-curriculum/tree/master/tu-python-libraries-rokem
