### 2020-07-29 #2407 @Xunius
@kbarnhart I got another comment regarding [installation](https://github.com/ihesp/IPART/issues/5). I'd like to propose a merge of the `netcdf4` branch to master, which would replace the `cdat` dependency with `netcdf4`. I already got some user feedbacks complaining about the `cdat` dependency, some about it not working well in google colab, some about it not supported in windows. I think this proposed change would make packaging to pip or conda-forge easier. I've already implemented the changes mentioned in the other comment [API documentation: undocumented parameters & methods](https://github.com/ihesp/IPART/issues/3), so we are not loosing the new corrections.
