### 2020-07-29 #2407 @rabernat
I am quite biased here, but as long as you are considering revamping your I/O routines, I would suggest you to look at xarray.

The requirement to operate on files period is a big limitation of this tool. What if the user's data are not in netCDF format. Other formats, like GRIB or Zarr, are common. (For example, there are nearly 1PB of [CMIP6 data in Google Cloud](https://cloud.google.com/blog/products/data-analytics/new-climate-model-data-now-google-public-datasets) in Zarr format.) Or if the user's data are not in files at all, but rather computed on the fly from some other source?

If you packages routines were able to accept xarray Dataset objects, you could effectively forget about dealing with files. Let the user open the files and pass an `xarray.Dataset` to your tool. This object should contain all raw numpy data, plus any metadata you need to process. This is the approach used by [xESMF](https://xesmf.readthedocs.io/en/latest/) (a regridding package), and it works very well.

We wrote a little bit about this concept here: http://pangeo.io/packages.html#guidelines-for-new-packages

I want to clarify that I don't think this sort of refactor is required to have your JOSS paper published. It's just an optional suggestion based on lots of experience with this sort of thing.
