### 2020-07-29 #2407 @kbarnhart
@Xunius I think that changing the dependency makes sense. I had hoped that some of my early feedback in pre-review on installation instructions and packaging would have addressed some of these install issues. This is all to say that I strongly support @rabernat 's comments on ihesp/IPART/issues/5 to follow installation and practicing best practices.

I also favor xarray over netcdf4, though if you have reasons for one dependency over another, that is your decision to make.

Similarly, I've had good success with taking I/O out of the core of a package (e.g., a user is expected to provide a numpy array, or xarray.Dataset) rather than a file. A user can convert many common file formats into these data-structures in one or two lines of code, so I don't think it places an unnecessary burden on users.

Regarding which things are and are not necessary to complete the JOSS process: I think that good packaging definitely is necessary and I would agree with @rabernat that refactoring the I/O is not necessary. You may choose to do it because it makes packaging easier, but that is up to you.

If additional extensive discussion on this point is necessary, I may move the thread to an in-repo issue.

As always, let me know how I can help with questions/clarifications.
