### 2020-07-29 ihesp/IPART#6 @rabernat
Community Guidlines
This issue is in reference to my JOSS review (https://github.com/openjournals/joss-reviews/issues/2407)

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I didn't find this information in your docs. However, it does appear to be in your README.

> If you encounter problems or would like to help improve the code, please don't hesitate to fire up an issue or pull request.

This is a pretty terse contributor guide. You may wish to elaborate it a bit.
