+++ ipart/recon.py
@@ -1,380 +1,0 @@
-
-import numpy as np
-
-
-def rank_order(image):
-    """Return an image of the same shape where each pixel is the
-    index of the pixel value in the ascending order of the unique
-    values of `image`, aka the rank-order value.
-
-    Parameters
-    ----------
-    image: ndarray
-
-    Returns
-    -------
-    labels: ndarray of type np.uint32, of shape image.shape
-        New array where each pixel has the rank-order value of the
-        corresponding pixel in `image`. Pixel values are between 0 and
-        n - 1, where n is the number of distinct unique values in
-        `image`.
-
-    original_values: 1-D ndarray
-        Unique original values of `image`
-
-    Examples
-    --------
-    >>> a = np.array([[1, 4, 5], [4, 4, 1], [5, 1, 1]])
-    >>> a
-    array([[1, 4, 5],
-           [4, 4, 1],
-           [5, 1, 1]])
-    >>> rank_order(a)
-    (array([[0, 1, 2],
-           [1, 1, 0],
-           [2, 0, 0]], dtype=uint32), array([1, 4, 5]))
-    >>> b = np.array([-1., 2.5, 3.1, 2.5])
-    >>> rank_order(b)
-    (array([0, 1, 2, 1], dtype=uint32), array([-1. ,  2.5,  3.1]))
-    """
-    flat_image = image.ravel()
-    sort_order = flat_image.argsort().astype(np.uint32)
-    flat_image = flat_image[sort_order]
-    sort_rank = np.zeros_like(sort_order)
-    is_different = flat_image[:-1] != flat_image[1:]
-    np.cumsum(is_different, out=sort_rank[1:])
-    original_values = np.zeros((sort_rank[-1] + 1,), image.dtype)
-    original_values[0] = flat_image[0]
-    original_values[1:] = flat_image[1:][is_different]
-    int_image = np.zeros_like(sort_order)
-    int_image[sort_order] = sort_rank
-    return (int_image.reshape(image.shape), original_values)
-
-def reconstructionLSM(seed, mask, lsm, landocean, method='dilation', selem=None,
-        offset=None):
-    """Perform a morphological reconstruction of an image.
-
-    Morphological reconstruction by dilation is similar to basic morphological
-    dilation: high-intensity values will replace nearby low-intensity values.
-    The basic dilation operator, however, uses a structuring element to
-    determine how far a value in the input image can spread. In contrast,
-    reconstruction uses two images: a "seed" image, which specifies the values
-    that spread, and a "mask" image, which gives the maximum allowed value at
-    each pixel. The mask image, like the structuring element, limits the spread
-    of high-intensity values. Reconstruction by erosion is simply the inverse:
-    low-intensity values spread from the seed image and are limited by the mask
-    image, which represents the minimum allowed value.
-
-    Alternatively, you can think of reconstruction as a way to isolate the
-    connected regions of an image. For dilation, reconstruction connects
-    regions marked by local maxima in the seed image: neighboring pixels
-    less-than-or-equal-to those seeds are connected to the seeded region.
-    Local maxima with values larger than the seed image will get truncated to
-    the seed value.
-
-    Parameters
-    ----------
-    seed : ndarray
-        The seed image (a.k.a. marker image), which specifies the values that
-        are dilated or eroded.
-    mask : ndarray
-        The maximum (dilation) / minimum (erosion) allowed value at each pixel.
-    method : {'dilation'|'erosion'}
-        Perform reconstruction by dilation or erosion. In dilation (or
-        erosion), the seed image is dilated (or eroded) until limited by the
-        mask image. For dilation, each seed value must be less than or equal
[…]
