### 2020-07-30 391f362
add numpy to setup.py
