### 2020-07-30 c9f5cdd
use setuptools.setup
