### 2020-07-30 ihesp/IPART#5 @Xunius
@rabernat Thanks for the comments. I ran into some difficulties addressing these issues.

I tried to upload to PyPI, but it seems that `basemap` is no longer install-able using `pip`:

```
(test) 😷:~$ pip3 install basemap
ERROR: Could not find a version that satisfies the requirement basemap (from versions: none)
ERROR: No matching distribution found for basemap
```
So it won't work as a dependency.

Then I tried building a conda package, but failed again. Previously I managed to `conda-build` a `cdat-lite` based version for py2, I don't know what changes I made that make the new attempts fail. The trackback starts from a line of `license='GPL-3` (see below), which is really confusing.

This is my `meta.yml`:

```
package:
    name: ipart
    version: '2.0'
source:
    git_url: https://github.com/ihesp/IPART.git
    git_tag: master
build:
    number: 0
requirements:
    build:
        - python
    run:
        - python 3.7*
        - numpy>=1.16
        - scipy
        - netcdf4
        - pandas
        - scikit-image
        - networkx
        - matplotlib
        - basemap
about:
    home: https://github.com/ihesp/IPART
    license: GPL-3.0
```

The `build.sh`:
```
#!/bin/bash
python setup.py install
```

Both `build.sh` and `meta.yml` are in a folder `ipart`, which also contains a `LICENSE` and a `bld.bat`.

The `conda-build ipart` outputs (starting from errors):

```
Installed $PREFIX/lib/python3.7/site-packages/ipart-2.0.2-py3.7.egg
Processing dependencies for ipart==2.0.2
Searching for basemap==1.2.0
Traceback (most recent call last):
  File "setup.py", line 37, in <module>
    license='GPL-3'
  File "/home/guangzhi/anaconda3/conda-bld/ipart_1596074610735/_h_env_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_pla/lib/python3.7/site-packages/setuptools/__init__.py", line 165, in setup
    return distutils.core.setup(**attrs)
  File "/home/guangzhi/anaconda3/conda-bld/ipart_1596074610735/_h_env_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_pla/lib/python3.7/distutils/core.py", line 148, in setup
    dist.run_commands()

<<<a number of similar lines starting "File "/home/guangzhi/anaconda3/conda-bld/ipart_1596074610735">>>

  File "/home/guangzhi/anaconda3/conda-bld/ipart_1596074610735/_h_env_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_pla/lib/python3.7/site-packages/setuptools/package_index.py", line 327, in process_url
    "Be sure to add all dependencies in the meta.yaml  url=%s" % url)
RuntimeError: Setuptools downloading is disabled in conda build. Be sure to add all dependencies in the meta.yaml  url=https://pypi.org/simple/basemap/
Traceback (most recent call last):
  File "/home/guangzhi/anaconda3/bin/conda-build", line 11, in <module>
    sys.exit(main())
  File "/home/guangzhi/anaconda3/lib/python3.7/site-packages/conda_build/cli/main_build.py", line 469, in main
    execute(sys.argv[1:])
  File "/home/guangzhi/anaconda3/lib/python3.7/site-packages/conda_build/cli/main_build.py", line 460, in execute
    verify=args.verify, variants=args.variants)
  File "/home/guangzhi/anaconda3/lib/python3.7/site-packages/conda_build/api.py", line 209, in build
    notest=notest, need_source_download=need_source_download, variants=variants)
  File "/home/guangzhi/anaconda3/lib/python3.7/site-packages/conda_build/build.py", line 2344, in build_tree
    notest=notest,
  File "/home/guangzhi/anaconda3/lib/python3.7/site-packages/conda_build/build.py", line 1492, in build
    cwd=src_dir, stats=build_stats)
  File "/home/guangzhi/anaconda3/lib/python3.7/site-packages/conda_build/utils.py", line 398, in check_call_env
    return _func_defaulting_env_to_os_environ('call', *popenargs, **kwargs)
  File "/home/guangzhi/anaconda3/lib/python3.7/site-packages/conda_build/utils.py", line 378, in _func_defaulting_env_to_os_environ
    raise subprocess.CalledProcessError(proc.returncode, _args)
subprocess.CalledProcessError: Command '['/bin/bash', '-o', 'errexit', '/home/guangzhi/anaconda3/conda-bld/ipart_1596074610735/work/conda_build.sh']' returned non-zero exit status 1.
```
I think I've added all dependencies in both the [setup.py](https://github.com/ihesp/IPART/blob/master/setup.py) and `meta.yml`. Do you spot anything wrong? What is missing?
