### 2020-07-30 ihesp/IPART#5 @rabernat
For building a conda package, the only way to go is to use conda forge: https://conda-forge.org/#add_recipe
Don't waste time trying to run conda-build yourself.

> it seems that `basemap` is no longer install-able using `pip`

Basemap is no longer maintained. [Cartopy](https://scitools.org.uk/cartopy/docs/latest/) has replaced it.

But more generally, why does your package need to depend on a mapping tool at all? Or any other visualization package? Of course one needs visualization to plot the results, but that's not the same thing as a dependency.
