### 2020-08-02 d27ebb2
change install guide to include conda-forge install
