### 2020-08-02 def1379
rephrase contributing guide in README and doc
