+++ README.md
@@ -125,7 +125,12 @@ Make algorithms zonally cyclic.
 
 
 
-## Contribution
+## Contributing
 
-If you encounter problems or would like to help improve the code, please don't
-hesitate to fire up an issue or pull request.
+We welcome contributions from the community. Please create a fork of the
+project on GitHub and use a pull request to propose your changes. We strongly encourage creating
+an issue before starting to work on major changes, to discuss these changes first.
+
+## Getting help
+
+Please post issues on the project GitHub page.
