### 2020-08-02 ihesp/IPART#6 @Xunius
Please checkout [this commit](https://github.com/ihesp/IPART/commit/def1379d471d51bb9faa567fa40018f1f58fed8b), I copied from https://github.com/neurohackademy/nh2020-curriculum/blob/master/tu-python-libraries-rokem/03-setup-file.md
