### 2020-08-02 #2407 @sadielbartholomew
>  I'd like to propose a merge of the netcdf4 branch to master, which would replace the cdat dependency with netcdf4. I already got some user feedbacks complaining about the cdat dependency, some about it not working well in google colab, some about it not supported in windows. I think this proposed change would make packaging to pip or conda-forge easier

> > @Xunius I think that changing the dependency makes sense.

I second/third that this would be a good idea. Further evidence that this is causing installation difficulty being that the (immediate) aspect that seemed to prevent ``python setup.py develop`` from working (before it was cut out for that reason in ihesp/IPART#4) was ``cdutil``:

```
...
No local packages or working download links found for cdutil
error: Could not find suitable distribution for Requirement.parse('cdutil')
```

which appears to be [a CDAT-related tool](https://github.com/CDAT/cdutil). However, that is just an illustration since I do also agree with the sentiment in ihesp/IPART#5 that you should aim to support a standard installation via ``pip install`` and/or ``conda install`` as that development install command is designed for those wanting to make changes to the codecase rather than use it as a package.

****

Regarding the I/O points, whilst I do agree that:

* xarray is a fantastic library that is becoming common in the field, such that it would be ideal to also support xarray data structures directly;
* generally it would be ideal also to take non-file inputs;

NetCDF (as stated in the current paper) is (surely) still the most widely used format for geoscience data. So it is most likely the user has their data contained in netCDF file(s), but in the minority of cases they do not, I think the central point is, as @kbarnhart notes:

> A user can convert many common file formats into these data-structures in one or two lines of code

but also vice versa, such that in answer to:

> What if the user's data are not in netCDF format?

ultimately it of course becomes the case of manual conversion and/or use of numerous tools that can straightforwardly convert between formats, before use of the IPART API. Slightly more work naturally , but I don't think it is a massive barrier to usage of, nor a major deficiency, to IPART, more a 'nice-to-have'. (Saying that, I sense I have a lot less experience than everyone here in the field, so as a caveat please don't weight my opinion too highly :smile: .)
