### 2020-08-02 ihesp/IPART#5 @Xunius
I think I've managed to get a PR merged to conda-forge, and I just tested that it is installable using `conda install -c conda-forge ipart` in my Linux box.
Please also checkout [this commit](https://github.com/ihesp/IPART/commit/d27ebb2a66c0ce9e3f0e1e29bc7cf55593bfef51) where I updated the installation guides.
