### 2020-08-02 #2407 @sadielbartholomew
Continuing on the topic of improvements that are not compulsory towards acceptance in the paper given the open criteria, but would be good to think about going forwards, for good practice with metadata I suggest making more use of the [CF Conventions](https://cfconventions.org/) (the [recommended standard for netCDF](https://www.unidata.ucar.edu/software/netcdf/conventions.html)), namely as described in the three points below.

1. Increasing the compliance of the datasets included in the repository to the CF Conventions, especially those under the ``notebooks`` directory which users may interact with if they try out IPART with the provided Notebooks. Notably both ``uflux_s_6_1984_Jan.nc`` & ``vflux_s_6_1984_Jan.nc`` provided there are marked by global attribute as being CF-compliant to CF 1.6:

    ```
    :Conventions = "CF-1.6" ;
    ```
    which is okay (relative to the ideal, latest version, 1.8), but immediately I see improvements in compliance that could be made.

    For example, the variable & dimensions are all described by a ``long_name`` attribute, where use of a ``standard_name`` attribute is preferable as each is unambiguous (see e.g. [here](http://cfconventions.org/cf-conventions/cf-conventions.html#long-name)). The time, lat & lon coordinates can take standard names of the same identifier as currently used for the long name, and from a quick search on the [names table](https://cfconventions.org/Data/cf-standard-names/73/build/cf-standard-name-table.html) for "eastward" AND "water" I think the data itself with ``long_name=Vertical integral of eastward water vapour flux`` and units ``kg m**-1 s**-1`` could probably be assigned a standard name of ``eastward_atmosphere_water_vapor_transport_across_unit_distance``, or similar.

2. Rephrasing aspects of the 'Data preparation' section of the documentation in terms of terminology from the CF Conventions. For instance, instead of stating there that:

     > Source data are the u- and v- components of the vertically integrated vapor fluxes, in a rectangular grid.

    you could explicitly state the standard names of data variables which would be applicable, e.g. something similar to ``northward_...`` and ``eastward_atmosphere_water_vapor_transport_across_unit_distance`` and maybe link to the definition of all grids which may be considered rectangular, for clarity. This would make it crystal clear whether a user's dataset(s) may be appropriate for processing by IPART.

3. Making code changes to accommodate conventions. In particular, the required ordering of dimensions for IPART seemingly go against conventions given that the 'Data preparation' section states that: "the user is responsible for making sure that the data are saved in the following rank order: (time, level, latitude, longitude)" but as conveyed in [this section](http://cfconventions.org/cf-conventions/cf-conventions.html#coards-relationship)  "The CF convention places no rigid restrictions on the order of dimensions, however we encourage data producers to make the extra effort to stay within the COARDS standard order" where "COARDS restricts the axis (equivalently dimension) ordering to be longitude, latitude, vertical, and time".

    So, you are advocating that users define data dimensions in the inverse order to that recommended. To make IPART more immediately accessible, you could amend your code so that it accepts the outlined conventional order, rather than the inverse.
