### 2020-08-03 #2407 @kbarnhart
@sadielbartholomew thanks for the update and these comments. No worries about timeframe, you and @rabernat have been very punctual with your reviews (JOSS also is striving to be realistic and understanding about how COVID-19 has impacted reviewer and author availability).
