### 2020-08-04 3a4a6fc
change CDAT->netcdf4 in paper
