+++ joss/paper.bib
@@ -1,3 +1,14 @@
+@software{netcdf,
+  author       = {Unidata},
+  title        = {Network Common Data Form (net{CDF}) version 4.7.4 [software]},
+  month        = April,
+  year         = 2020,
+  publisher    = {NCAR/Unidata Program Center},
+  version      = {v4.7.4},
+  doi          = {https://doi.org/10.5065/D6H70CW6}
+  url          = {https://doi.org/10.5065/D6H70CW6}
+}
+
 @software{CDAT,
   author       = {Charles Doutriaux and
                   Denis Nadeau and
@@ -17,7 +28,7 @@
 
 @ARTICLE{numpy,
   author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},
-  journal={Computing in Science   Engineering},
+  journal={Computing in Science Engineering},
   title={The NumPy Array: A Structure for Efficient Numerical Computation},
   year={2011},
   volume={13},
