### 2020-08-04 a335140
change mock imports in sphnix docs after cdat->netcdf4 change
