+++ docs/conf.py
@@ -63,5 +63,5 @@ napoleon_numpy_docstring = False
 napoleon_use_rtype = False
 napoleon_include_init_with_doc = True
 
-autodoc_mock_imports = ['numpy', 'scipy', 'cdms2', 'MV2', 'cdutil', 'matplotlib',
-        'pandas', 'networkx', 'skimage', 'basemap', 'genutil', 'mpl_toolkits']
+autodoc_mock_imports = ['numpy', 'scipy', 'netCDF4', 'matplotlib',
+        'pandas', 'networkx', 'skimage', 'basemap', 'mpl_toolkits']
