### 2020-08-04 #2407 @Xunius
> > I plan, and will strive, to have my review completed by the end of this week i.e. the end of July.
>
> Sorry, I am running a bit later than this, but I promise to complete my review in the next few days.
>
> Just to note regarding some newly-ticked review points:
>
>     * I have checked off 'Community guidelines' as they are now clear in both the docs & README after [ihesp/IPART@def1379](https://github.com/ihesp/IPART/commit/def1379d471d51bb9faa567fa40018f1f58fed8b);
>
>     * I am checking off the 'Performance' item since I cannot find any performance claims.
>
>
> > I'd like to propose a merge of the netcdf4 branch to master, which would replace the cdat dependency with netcdf4
>
> @Xunius I notice you have now merged in your `netcdf4` branch commits. This is good, & I will check the installation again with this new dependency set.
>
> I notice you have added more to the documentation too, which is also good, however in case you are not aware I have noticed that the API documentation pages (e.g. ['Documentation page for AR_detector.py'](https://ipart.readthedocs.io/en/latest/AR_detector.html)) have become blank after your latest changes, so please make sure these are re-populated.
>
> Just as a reminder, you should update the 'External libraries used' section of the JOSS paper content to reflect the `cdat` -> `netcdf4` change. On that note, I have noticed some points in the paper itself which are not clear, and some minor formatting issues, so wll put in a PR in a moment to suggest how these can be resolved.

* I have made the change from `CDAT` to `netCDF` in joss paper in [this commit](3a4a6fc303b65468f78051ae82d864ec1d7c5e65). I guess we should call the bot to rebuild the PDF?

* Those missing API documentation pages are not back: https://ipart.readthedocs.io/en/latest/AR_detector.html.
* Thanks for your PR regarding the joss paper fixes. I've merged it.
