### 2020-08-04 #2407 @kbarnhart
@Xunius its definitely fine to copy (with attribution) @sadielbartholomew 's comments into an in-repo issue. I would recommend doing that.

Also, in order for whedon the command must be the first part of the comment.
