### 2020-08-04 #2407 @sadielbartholomew
Thanks @Xunius.

> Thanks, I'll think about these. Do you mind if I copy your comments to a new issue in the IPART repo, just to make it easier for me get back to them later?

Please do. Sorry, maybe it would have been better for me to have put them in an Issue originally.

> Regarding Point 3. The major reason for using the time, lev, lat, lon ordering is that the reanalyses seem to follow this (e.g. the sample ERA-I data included in the notebooks, and the MERRA2 which the ARTMIP community tries to make a standard for reanslysis based AR detections). So I felt that it would feel a bit confusing the inputs and outputs are in different orderings.

Ah I see; since there's already a more specific standard you are following, it's of course best to leave it as-is. So long as you have actively thought about the best order, ideally abiding by some formal or informal community standard, I think it is fine.

> Or maybe it would be also good to somehow detect the input ordering (in a robust way) and apply the same ordering to the outputs. I'll have to think about this later.

This is certainly the best solution going forward though, I would say. Anything that minimises the amount of pre-processing a potential user of IPART would have to do on any dataset they might want to use, would be useful.
