### 2020-08-04 #2407 @sadielbartholomew
Possibly my PR on the text broke the formatting somehow. If so, my apologies. Though from the note on https://github.com/openjournals/joss-reviews/issues/2407#issuecomment-668609394 it appears to stem from the BibTeX/bibliography so it may be a different change. Either way, let me know if I can help to fix it.
