### 2020-08-04 925c793
fix paper.bib formatting
