+++ joss/paper.bib
@@ -5,7 +5,7 @@
   year         = 2020,
   publisher    = {NCAR/Unidata Program Center},
   version      = {v4.7.4},
-  doi          = {https://doi.org/10.5065/D6H70CW6}
+  doi          = {https://doi.org/10.5065/D6H70CW6},
   url          = {https://doi.org/10.5065/D6H70CW6}
 }
 
