### 2020-08-04 #2407 @Xunius
Should just be my incorrect bib file formatting.
