### 2020-08-05 ihesp/IPART#9 @Xunius
@sadielbartholomew I read about the `standard_name` in the links you gave and now I understand better about the differences between `long_name` and `standard_name`, I used to use the same string for both.

Regarding Point 2, the standard_names for the u- and v- flux components are indeed, as you said, `eastward_atmosphere_water_transport_across_unit_distance` and `northward_atmosphere_water_transport_across_unit_distance`.

Regarding Point 1, the `uflux_s_6_1984_Jan.nc` data in the notebooks folder are directly obtained from the [ERA-I reanalysis data center](https://apps.ecmwf.int/datasets/data/interim-full-daily/levtype=sfc/), I selected the desired variable, time step and domain etc.. and downloaded them. So they came without a `standard_name` attribute, maybe it is the grib to nc conversion in their data server that omits the attribute. I never noticed this before.

If I understand correctly, there exists a pre-defined, permissible list of `standard_name`s, so I can't just coin my new `standard_name`, like `numerical_label_for_atmospheric_river` or `northward_atmosphere_water_transport_across_unit_distance_THR_anomaly_component`, can I? In that case, do I just leave the attribute empty?
