### 2020-08-05 b8fdd47
standard_name and unit changes in notebooks, docs
Unit change: kg/m/s -> kg/(m \cdot s) or kg/(m*s)

Data-preparation.rst, AR_detector.py, 1_compute_IVT.ipynb,
2_compute_THR.ipynb, 3_detect_ARs.ipynb, Index.ipynb, compute_ivt.py,
detect_ARs.py, detect_ARs_generator_version.py:

give uflux, vflux their `standard_name`s in docs.
