### 2020-08-05 0ef1561
fix standard_name attributes
AR_detector.py:

    in getARData(), replace spaces with _ in "standard_name" for labels,
    angles, corssfluxes.

thr.py:

    add a helper func getAttrDict() to prepare attribute dict for output
    nc data.

    in THR(), rotatingTHR(), call getAttrDict() to prepare long_name,
    standard_name, units attrs.
