### 2020-08-06 ihesp/IPART#10 @sadielbartholomew
Explicitly outline target audience
The documentation makes the design purpose of IPART clear, but the target audience should also be described as a further component of the 'statement of need' criterion for inclusion of a project in JOSS.

Please, towards the openjournals/joss-reviews#2407 review, provide an outline of the intended users for IPART. This can be very brief, e.g. a single sentence. As an example, I had a look at some packages with published JOSS papers and found a good example [here for `geemap`](https://geemap.readthedocs.io/en/latest/readme.html#introduction), namely the paragraph starting (though a similar sentence alone would be sufficient):

> geemap is intended for students and researchers, who would like to utilize the Python ecosystem of diverse libraries and tools to explore Google Earth Engine.

Also, though I wouldn't count any of these as compulsory to satisfying the criteria above, I think it would be really useful to:

1. have the project description from the documentation, i.e:

    > IPART (Image-Processing based Atmospheric River Tracking) is a Python package for automated Atmospheric River (AR) detection, axis finding and AR tracking from gridded Integrated Vapor Transport (IVT) data, for instance Reanalysis datasets.

    (or perhaps even more from the 'Introduction section') at the top of the README file, as well as the start of the documentation. I believe it is common for people to check the README to understand the basics about a project. I certainly do that most of the time.
2. provide a link to the documentation from the 'About' section of the project repo (see for example the [`netcdf4-python` repo homepage](https://github.com/Unidata/netcdf4-python), so it can be accessed without having to go quite far down in the README to find the link.
