### 2020-08-06 ihesp/IPART#3 @sadielbartholomew
Thanks @Xunius, I have checked the status of the documentation again and you have addressed my feedback well with those new commits & others, so I can close this. I am happy to tick the 'Functionality documentation' criteria off the checklist, especially given the new detailed sections for each of the four core functionalities (the 'Perform the THR computation on IVT data' page, etc).
