### 2020-08-07 ihesp/IPART#10 @sadielbartholomew
I should also note that a description of the target audience is also required in the paper itself (see the 'statement of need' criterion under the 'software paper' sub-list). So please incorporate at least a brief part of the audience description into the paper, too. Thanks.
