### 2020-08-07 #2407 @sadielbartholomew
The references are very comprehensive & appropriately cited, which is great. @Xunius there are two small points (too minor for a dedicated Issue I think) I would like addressed before I tick the 'software paper -> references' item, though:

* two of the links seem to be broken, or at least they do not work for me right now:
    * Ralph2011 (https://github.com/ihesp/IPART/blob/0ef1561a7fe23db4723bb289ce07b1ae74a9a899/joss/paper.bib#L232-L248)
    * pandas (https://github.com/ihesp/IPART/blob/0ef1561a7fe23db4723bb289ce07b1ae74a9a899/joss/paper.bib#L73-L81)

     so please double check the information for those entries are correct. If they are, I will try them again in case it was temporary e.g. caused by some server downtime or similar.
* I was thinking there would likely be a DOI for the matplotlib reference and it seems there is, as stated [here](https://matplotlib.org/citing.html) in their official docs, so it would be good to add that to your bibliography entry.
