### 2020-08-07 75d82c1
add target audience
To address issue https://github.com/ihesp/IPART/issues/10.

README.md:
    add an Introduction subsection.

    add target audience.

    put Dependencies and Installation subsections after Documentation
    and Example use.

paper.md, paper.bib:

    add target audience.

    Cite Xu2020.
