+++ joss/paper.bib
@@ -667,3 +667,15 @@ year = {2019}
     volume = {11},
     year = {2018}
 }
+
+@article{Xu2020,
+AUTHOR = {Xu, Guangzhi and Ma, Xiaohui and Chang, Ping and Wang, Lin},
+TITLE = {A Comparison of Northern Hemisphere Atmospheric Rivers Detected by a New Image-Processing Based Method and Magnitude-Thresholding Based Methods},
+JOURNAL = {Atmosphere},
+VOLUME = {11},
+YEAR = {2020},
+NUMBER = {6},
+ARTICLE-NUMBER = {628},
+ISSN = {2073-4433},
+DOI = {10.3390/atmos11060628}
+}
