### 2020-08-07 9d6ea18
fix some citation links
in response to https://github.com/openjournals/joss-reviews/issues/2407#issuecomment-670261273.

Not all issues fixed:

* matplotlib ref: replaced with the one given [here](https://matplotlib.org/citing.html)
* pandas ref: added url of the home page https://pandas.pydata.org/.
* Ralph2011 ref: don't know what's wrong. The doi is correct, but the
  page is gone in the AMS site.
