+++ joss/paper.bib
@@ -60,14 +60,19 @@
       doi = {https://doi.org/10.1038/s41592-019-0686-2},
 }
 
-@ARTICLE{matplotlib,
-  author={J. D. {Hunter}},
-  journal={Computing in Science   Engineering},
-  title={Matplotlib: A 2D Graphics Environment},
-  year={2007},
-  volume={9},
-  number={3},
-  pages={90-95},
+@Article{matplotlib,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
 }
 
 @InProceedings{pandas,
@@ -77,7 +82,8 @@
   pages     = { 56 - 61 },
   year      = { 2010 },
   editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
-  doi       = { 10.25080/Majora-92bf1922-00a }
+  doi       = { 10.25080/Majora-92bf1922-00a },
+  url       = { https://pandas.pydata.org/ }
 }
 
 @article{scikit-image,
@@ -239,12 +245,7 @@ number = {4},
 pages = {1169-1189},
 year = {2011},
 doi = {10.1175/2010MWR3596.1},
-URL = {
-        https://doi.org/10.1175/2010MWR3596.1
-},
-eprint = {
-        https://doi.org/10.1175/2010MWR3596.1
-}
+URL = {https://doi.org/10.1175/2010MWR3596.1},
 }
 
 @article{Guan2015,
