### 2020-08-07 5227644
add target audience to docs/index.rst
