### 2020-08-07 ihesp/IPART#10 @Xunius
@sadielbartholomew Thanks.

Please checkout these 2 commits:
* https://github.com/ihesp/IPART/commit/75d82c16ff5cd452ab06a1b85431f4371d385bc9
* https://github.com/ihesp/IPART/commit/52276448b72a1bf0da69ffe09716d596d5eee08b
where I added target audience to the README.md, documentation and JOSS paper.

The README.md is also added a new Introduction subsection. I changed the order of subsections in README.md a bit, so that Dependencies and Installation appear later.

The documentation link is added to the *About* section of the repo.

The JOSS paper has a minor change to reflect that our paper submitted to Atmosphere has been published.
