### 2020-08-07 #2407 @Xunius
@sadielbartholomew Please checkout [this commit](https://github.com/ihesp/IPART/commit/9d6ea1887cefd7e95078756542c53dbbaf61c6d5):

* I used the citation of matplotlib you provided.
* Added the `url` field to the pandas reference.
* Regarding the `Ralph2011` citation. I double checked the DOI (10.1175/2010MWR3596.1) which seems to be correct, see [here](https://connect.unavco.org/display/pub912261) and [the paper PDF](http://pdfs.semanticscholar.org/f06a/0ecc6b4deec7d298b2a5077d68466c58af78.pdf). I also tried searching the DOI in https://www.doi.org/, it does lead to a broken page. Maybe there is something wrong in the doi parsing, I don't know.
