### 2020-08-07 #2407 @sadielbartholomew
Hi @kbarnhart, do you have any advice on procedure regarding citations with links that are correct but don't seem to work for some unknown reason? Is it okay to keep those in the paper, or should the URL be removed for safety in case in generally does not work as opposed to temporarily being down:

> Regarding the Ralph2011 citation. I double checked the DOI (10.1175/2010MWR3596.1) which seems to be correct, see here and the paper PDF. I also tried searching the DOI in https://www.doi.org/, it does lead to a broken page. Maybe there is something wrong in the doi parsing, I don't know.

Other than this query, I am very happy with the references so can check that item off the review list.
