### 2020-08-07 #2407 @kbarnhart
@sadielbartholomew @Xunius It does appear that the DOI link is broken on the Monthly Weather Review site. I'll consult with the JOSS editorial board about what the best approach to handling this is.
