### 2020-08-11 #2407 @kbarnhart
A follow up regarding what to do about the Ralph et al. (2011) DOI. Based on consultation with the rest of the JOSS editorial board, we think the best thing to do is nothing. The citation information provided in the .bib file is correct and the paper is real. I don't know exactly why the DOI doesn't work, but it is clearly the correct DOI.
