### 2020-08-12 20b5f34
fix some dois
add doi to numpy.
del https://doi.org/ prefix for netcdf, scipy, matplotlib.
del wilks2011, not used.
