+++ joss/paper.bib
@@ -5,7 +5,7 @@
   year         = 2020,
   publisher    = {NCAR/Unidata Program Center},
   version      = {v4.7.4},
-  doi          = {https://doi.org/10.5065/D6H70CW6},
+  doi          = {10.5065/D6H70CW6},
   url          = {https://doi.org/10.5065/D6H70CW6}
 }
 
@@ -34,6 +34,7 @@
   volume={13},
   number={2},
   pages={22-30},
+  doi={10.1109/mcse.2011.37},
 }
 
 @ARTICLE{scipy,
@@ -57,7 +58,7 @@
       volume={17},
       pages={261--272},
       adsurl = {https://rdcu.be/b08Wh},
-      doi = {https://doi.org/10.1038/s41592-019-0686-2},
+      doi = {10.1038/s41592-019-0686-2},
 }
 
 @Article{matplotlib,
@@ -217,7 +218,7 @@ number = "3",
 pages = "244 - 256",
 year = "1972",
 issn = "0146-664X",
-doi = "https://doi.org/10.1016/S0146-664X(72)80017-0",
+doi = "10.1016/S0146-664X(72)80017-0",
 url = "http://www.sciencedirect.com/science/article/pii/S0146664X72800170",
 author = "Urs Ramer",
 }
@@ -512,14 +513,6 @@ ISSN={1057-7149},
     title = {Genesis, Pathways, and Terminations of Intense Global Water Vapor Transport in Association with Large-Scale Climate Patterns},
     volume = {44},
     year = {2017}
-}
-
-@book{wilks2011,
-author = {Wilks, Daniel S},
-publisher = {Academic press},
-title = {{Statistical methods in the atmospheric sciences}},
-volume = {100},
-year = {2011}
 }
 
 @misc{era5,
