### 2020-08-12 #2407 @Xunius
[this commit](https://github.com/ihesp/IPART/commit/20b5f349c100f9577b494c8e0cf995fbede80a8d) fixed the 3 invalid DOIs, added the missing DOI (10.1109/mcse.2011.37) for numpy, and deleted the record for _Statistical methods in the atmospheric sciences_ which is not referenced in paper.
