### 2020-08-12 #2407 @sadielbartholomew
> A follow up regarding what to do about the Ralph et al. (2011) DOI. Based on consultation with the rest of the JOSS editorial board, we think the best thing to do is nothing.

Thanks for asking about this and reporting back, @kbarnhart. That sounds like a sensible decision to me.
