### 2020-08-12 fb6a9fe
Force BibTeX to leave capitalisation on networkx reference
