### 2020-08-12 #2407 @sadielbartholomew
Great, thanks @Xunius the references are all good.

Though, if I may be pedantic, I've noticed that the reference for ``networkx`` that is correct in the ``.bib`` is converted to lower case by BibTeX and therefore on the rendered paper it goes from having the correct capitalisation, as in:

https://github.com/ihesp/IPART/blob/20b5f349c100f9577b494c8e0cf995fbede80a8d/joss/paper.bib#L111

to being all lower-case ("Proceedings of the 7th python in science conference"). It seems minor, but Python should strictly always be capitalised so we should fix this. I already played around to see what might be going on for my own understanding, and know how to fix it, so I put in a small PR, as referenced above.

I'm checking off the references component. I hope to finish my review by Friday.
