### 2020-08-24 ihesp/IPART#5 @kbarnhart
@Xunius these changes are a great step towards making IPART much easier to install. 👍

Looking at your [meta.yaml](https://github.com/conda-forge/ipart-feedstock/blob/master/recipe/meta.yaml) for conda-forge, it looks like no tests run as part of making the conda-forge build. Recipes should include at least some tests (see, for example, [this part](https://conda-forge.org/docs/maintainer/adding_pkgs.html#test) of the conda forge documentation).

Your tests take about 9 seconds to run, so I would recommend having them all run.

One other minor comment:
- In the Readme section on [tests](https://github.com/ihesp/IPART#tests) I would recommend making clear that running the tests will only work if one has done a source code install.
