### 2020-08-24 #2407 @kbarnhart
@Xunius thanks for checking in. The JOSS review process can sometimes be a bit more iterative than a traditional journal. For example, if there is an issue that prevents a reviewer from completing their review (e.g., installation issues), then they may start the review, make an issue, wait for the authors address it, and then the reviewers can resume their review.

It appears that there are a few outstanding issues

_necessary:_

1. Installation instructions https://github.com/ihesp/IPART/issues/5
2. Community guidelines https://github.com/ihesp/IPART/issues/6

_nice but not necessary:_

3. netcdf conventions https://github.com/ihesp/IPART/issues/9

As best as I can tell, you have made great progress in addressing (1) and (2). I've made a small note regarding incorporating tests into the conda-forge meta.yaml in the relevant issue.

As best as I can tell, once that is addressed, you will have addressed (1) and (2), and then it would be appropriate to ping me and the reviewers here to let us know that it makes sense for the review to proceed.

Let me know if you have any further questions or concerns.
