### 2020-08-24 ihesp/IPART#9 @kbarnhart
@Xunius a couple of follow on points about this:

1. I don't think that you need to do anything to files downloaded directly from the ERA-1 reanalysis data center as its not your role to make that file CF compliant.

2. I would _strongly_ recommend that you relax the requirements associated with dimension ordering and replace them with an expectation that specific dimensions are present (and named correctly). Netcdf and xarray both allow access to the name metadata. This will allow you to support both datasets that comply with the COARDS standard order and datasets such as the the ERA-I datacenter downloads which provide netcdfs in (time, lat, lon) order.

3. _Regarding standard names_: You are correct that you can't create a new standard name (e.g., something listed in the CF standard name table) on your own. My experience is coming up with these names is often challenging and a real art. My recommendation is that you use a descriptive name that follows the style of the standard names for the "standard_name" attribute field. Also take advantage of the "long name", "units", and "_FillValue" fields in order to complete the description.  As your example above shows, being descriptive often means making very very long names.

Also 👍 👍 to @sadielbartholomew for such a thorough comment on CF-compliance and standard names. 🚀
