### 2020-08-24 ihesp/IPART#5 @Xunius
@kbarnhart So if I understood you correctly, the suggested change is to add something like the following to the `meta.yaml`?

```
test:
  imports:
    - ipart
    - ipart.utils
  source_files:
    - tests
  commands:
    - python -m unittest discover -s tests
```
If so, is the README change still required?
