### 2020-08-25 ihesp/IPART#5 @kbarnhart
@Xunius, based on my understanding of IPART and conda-forge, what you've described above would be the appropriate change. With the test commands you then ensure that the packages built by conda-forge pass the tests before they are distributed.

I would still recommend making the README change because a user who installed with conda-forge would not be able to run `python -m unittest discover -s tests`. I've had users get confused that their install was broken when, in fact, they had just installed the binaries and shouldn't have been expected to run any tests.
