### 2020-08-26 dcfad9e
README change about tests
See https://github.com/ihesp/IPART/issues/5#issuecomment-679213870
