### 2020-08-26 d40d6dd
adding matplotlib.use('Agg') to test scripts
conda build tests failed with error:
_tkinter.TclError: no display name and no $DISPLAY environment variable
