+++ tests/test_thr.py
@@ -7,6 +7,10 @@ from __future__ import print_function
 import os
 import unittest
 import numpy as np
+# try fixing the issue of
+# _tkinter.TclError: no display name and no $DISPLAY environment variable
+import matplotlib
+matplotlib.use('Agg')
 from ipart.utils import funcs
 from ipart.thr import THR
 
