### 2020-08-26 ihesp/IPART#5 @Xunius
@kbarnhart After a bit of trial-n-error, I added unittest to the meta.yml recipe and [merged the PR](https://github.com/conda-forge/ipart-feedstock/pull/2).

Also edited README in [this commit](https://github.com/ihesp/IPART/commit/dcfad9ed4208c4551bba40928f64b940c07bda66)
