### 2020-08-26 ihesp/IPART#5 @kbarnhart
@Xunius thanks for making that change to the conda-forge recipe.

As this is @rabernat 's issue and is a bit more substantial than #6  I'm going to let him chime in on whether he thinks the issues he has raised have been addressed.

I want to follow up with some guidance from JOSS regarding dependencies, and in particular deprecated dependencies. It appears that basemap is a listed dependency because it is used to make some built in result plots. I consider whether or not these sort of functions are in or out of the package as a point of developer preference and will not express an editorial opinion on it here.

However, as @rabernat points out, basemap was deprecated ([quite a while ago](https://matplotlib.org/basemap/users/intro.html#cartopy-new-management-and-eol-announcement)) and replaced with cartopy.

I consulted with the rest of the JOSS editorial board regarding this case: Unless you can make a strong case that the deprecated dependency cannot be replaced by the new package or it would be a substantial burden to replace it, then you must replace it as part of the review. The JOSS :_ formal peer review process that is designed to improve the quality of the software submitted_. https://joss.theoj.org/about

Basemap is used relatively infrequently in ipart, and thus I do not think it is an undue burden to have you replace its use with cartopy.
