### 2020-08-27 f0b1572
replacing basemap with cartopy, core part
AR_detector.py:

    in plotAR(), replace bmap input arg with lonax, which is used when
    calling funcs.breakCurveAtEdge(). The plotting of contours and axes
    are using the `transform=ccrs.PlateCarree()` kwarg.

AR_tracer.py:

    in plotAR(), set domain extents, plot coastlines and AR axes using
    cartopy.

plot.py:

    add Plot2Cartopy class, mod from Plot2Basemap. Add a new plot2()
    func to work with Plot2Cartopy.

detect_ARs.py and detect_ARs_generator_version.py:

    use new plot2() and plotAR() signature.

trace_ARs.py:

    use new p
[…]
