+++ ipart/AR_tracer.py
@@ -12,10 +12,11 @@ import copy
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
-from mpl_toolkits.basemap import Basemap
+#from mpl_toolkits.basemap import Basemap
+import cartopy.crs as ccrs
 import matplotlib.patches as patches
 
-from ipart.utils import funcs, plot
+from ipart.utils import funcs
 
 
 #######################################################################
@@ -284,7 +285,7 @@ def plotHD(y1,x1,y2,x2,timelabel=None,linkflag='',ax=None,show=True):
 
     return
 
-def plotAR(arlist,latax,lonax,full=False,ax=None,label=None,linestyle='solid',
+def plotAR(arlist,latax,lonax,ax,full=False,label=None,linestyle='solid',
         marker=None):
     '''Plot AR tracks
 
@@ -309,22 +310,25 @@ def plotAR(arlist,latax,lonax,full=False,ax=None,label=None,linestyle='solid',
 
     if ax is None:
         figure=plt.figure(figsize=(12,6),dpi=100)
-        ax=figure.add_subplot(111)
+        ax=figure.add_subplot(111,projection=ccrs.PlateCarree())
         isshow=True
 
-    ax.patch.set_color('0.7')
+    lonax=np.array(lonax)
+    lon0=lonax[len(lonax)//2]
+    ax.projection=ccrs.PlateCarree(central_longitude=lon0)
 
     if not isinstance(arlist,(tuple,list)):
         arlist=[arlist,]
 
-    bmap=Basemap(projection='cyl',
-            llcrnrlat=latax[0],llcrnrlon=lonax[0],
-            urcrnrlat=latax[-1],urcrnrlon=lonax[-1],
-            ax=ax,fix_aspect=False)
+    #bmap=Basemap(projection='cyl',
+            #llcrnrlat=latax[0],llcrnrlon=lonax[0],
+            #urcrnrlat=latax[-1],urcrnrlon=lonax[-1],
+            #ax=ax,fix_aspect=False)
 
-    bmap.drawcoastlines(linewidth=0.5,linestyle='solid',color='k',\
-        antialiased=True)
+    #bmap.drawcoastlines(linewidth=0.5,linestyle='solid',color='k',\
+            #antialiased=True)
 
+    '''
     lon_labels=np.array(plot.mkscale(lonax[0],lonax[-1],15,1))
     idx=np.where((lon_labels>=lonax[0]) & (lon_labels<=lonax[-1]))
     lon_labels=np.array(lon_labels)[idx]
@@ -332,14 +336,21 @@ def plotAR(arlist,latax,lonax,full=False,ax=None,label=None,linestyle='solid',
     lat_labels=np.array(plot.mkscale(latax[0],latax[-1],15,1))
     idx=np.where((lat_labels>=latax[0]) & (lat_labels<=latax[-1]))
     lat_labels=np.array(lat_labels)[idx]
+    '''
 
-    parallels=[1,1,0,0]
-    meridians=[0,0,0,1]
+    ax.patch.set_color('0.7')
+    ax.set_extent([lonax[0], lonax[-1], latax[0], latax[-1]],
+            crs=ccrs.PlateCarree())
+    ax.coastlines()
+    ax.gridlines(draw_labels=True, zorder=-1, color='0.7')
 
-    bmap.drawparallels(lat_labels,labels=parallels,linewidth=0,\
-            labelstyle='+/-',fontsize=14)
-    bmap.drawmeridians(lon_labels,labels=meridians,linewidth=0,\
-            labelstyle='+/-',fontsize=14)
+    #parallels=[1,1,0,0]
+    #meridians=[0,0,0,1]
+
+    #bmap.drawparallels(lat_labels,labels=parallels,linewidth=0,\
[…]
