+++ ipart/utils/plot.py
@@ -7,6 +7,7 @@ Update time: 2020-07-22 09:27:30.
 #--------Import modules--------------
 import numpy
 import matplotlib.pyplot as plt
+import cartopy.crs as ccrs
 from . import funcs as functions
 
 DEFAULT_CMAP=plt.cm.PRGn
@@ -1649,8 +1650,194 @@ class Plot2Basemap(Plot2D):
 
         return cbar
 
+class Plot2Cartopy(Plot2D):
+    def __init__(self,var,method,ax,legend='global',title=None,\
+            xarray=None,yarray=None,\
+            latlon=True,latlongrid=False,fill_color='0.8',
+            legend_ori='horizontal',clean=False,fix_aspect=False):
 
-def plot2(var,method,ax=None,legend='global',
+        Plot2D.__init__(self,var,method,ax=ax,\
+                xarray=xarray,yarray=yarray,\
+                title=title,latlon=latlon,\
+                latlongrid=latlongrid,legend=legend,
+                legend_ori=legend_ori,clean=clean)
+
+        self.projection=ax.projection
+        self.fill_color=fill_color
+        self.fix_aspect=fix_aspect
+
+    def getGrid(self):
+
+        try:
+            latax=self.var.getLatitude()[:]
+            lonax=self.var.getLongitude()[:]
+            lons,lats=numpy.meshgrid(lonax,latax)
+        except:
+            lonax,latax,lons,lats=super(Plot2Cartopy,self).getGrid()
+
+        return lonax,latax,lons,lats
+
+
+    def _plot(self):
+
+        #-------------Plot according to method-------------
+        #-------------------Contour fill/line-------------------
+        bmap=self.ax
+        trans=ccrs.PlateCarree()
+
+        if self.method.method in ['isofill','isoline']:
+
+            if self.method.ext_1 is False and self.method.ext_2 is False:
+                extend='neither'
+            elif self.method.ext_1 is True and self.method.ext_2 is False:
+                extend='min'
+            elif self.method.ext_1 is False and self.method.ext_2 is True:
+                extend='max'
+            else:
+                extend='both'
+
+            if self.method.method=='isofill':
+                cs=bmap.contourf(self.lons,self.lats,self.var,self.method.levels,latlon=True,\
+                        cmap=self.method.cmap,extend=extend,hatch='/',
+                        transform=trans)
+
+            elif self.method.method=='isoline':
+                if self.method.color is not None:
+                    colors=[self.method.color]*len(self.method.levels)
+                    cs=bmap.contour(self.lons,self.lats,self.var,self.method.levels,latlon=True,\
+                            colors=colors,extend=extend,
+                            linewidth=self.method.linewidth,
+                            alpha=self.method.alpha,
+                            transform=trans)
+                else:
+                    if self.method.black:
+                        colors=['k']*len(self.method.levels)
+                        cs=bmap.contour(self.lons,self.lats,self.var,self.method.levels,latlon=True,\
+                                colors=colors,extend=extend,
+                                linewidth=self.method.linewidth,
+                                alpha=self.method.alpha,
+                                transform=trans)
+                    else:
+                        cs=bmap.contour(self.lons,self.lats,self.var,self.method.levels,latlon=True,\
+                                cmap=self.method.cmap,extend=extend,
+                                alpha=self.method.alpha,
+                                transform=trans)
+
+                #-----------------Set line styles-----------------
+                # for some reason it's not giving me dashed line for negatives.
[…]
