### 2020-08-28 ef281c5
doc change regarding cartopy. notebook change
