+++ ipart/utils/plot.py
@@ -1697,14 +1697,16 @@ class Plot2Cartopy(Plot2D):
                 extend='both'
 
             if self.method.method=='isofill':
-                cs=bmap.contourf(self.lons,self.lats,self.var,self.method.levels,latlon=True,\
-                        cmap=self.method.cmap,extend=extend,hatch='/',
+                cs=bmap.contourf(self.lons,self.lats,self.var,
+                        self.method.levels,
+                        cmap=self.method.cmap,extend=extend,
                         transform=trans)
 
             elif self.method.method=='isoline':
                 if self.method.color is not None:
                     colors=[self.method.color]*len(self.method.levels)
-                    cs=bmap.contour(self.lons,self.lats,self.var,self.method.levels,latlon=True,\
+                    cs=bmap.contour(self.lons,self.lats,self.var,
+                            self.method.levels,
                             colors=colors,extend=extend,
                             linewidth=self.method.linewidth,
                             alpha=self.method.alpha,
@@ -1712,13 +1714,15 @@ class Plot2Cartopy(Plot2D):
                 else:
                     if self.method.black:
                         colors=['k']*len(self.method.levels)
-                        cs=bmap.contour(self.lons,self.lats,self.var,self.method.levels,latlon=True,\
+                        cs=bmap.contour(self.lons,self.lats,self.var,
+                                self.method.levels,
                                 colors=colors,extend=extend,
                                 linewidth=self.method.linewidth,
                                 alpha=self.method.alpha,
                                 transform=trans)
                     else:
-                        cs=bmap.contour(self.lons,self.lats,self.var,self.method.levels,latlon=True,\
+                        cs=bmap.contour(self.lons,self.lats,self.var,
+                                self.method.levels,
                                 cmap=self.method.cmap,extend=extend,
                                 alpha=self.method.alpha,
                                 transform=trans)
@@ -1757,13 +1761,14 @@ class Plot2Cartopy(Plot2D):
         elif self.method.method == 'boxfill':
 
             cs=bmap.imshow(self.var,cmap=self.method.cmap,
-                    vmin=self.method.vmin,vmax=self.method.vmax,interpolation='nearest',
+                    vmin=self.method.vmin,vmax=self.method.vmax,
+                    interpolation='nearest',
                     transform=trans)
 
         #-------------------Pcolor fill-------------------
         elif self.method.method == 'pcolor':
 
-            cs=bmap.pcolormesh(self.lons,self.lats,self.var,latlon=True,\
+            cs=bmap.pcolormesh(self.lons,self.lats,self.var,
                     cmap=self.method.cmap,vmin=self.method.levels[0],\
                     vmax=self.method.levels[-1],
                     transform=trans)
@@ -1775,13 +1780,14 @@ class Plot2Cartopy(Plot2D):
                 nlevel=1
             else:
                 nlevel=3
-            cs=bmap.contourf(self.lons,self.lats,self.var,nlevel,latlon=True,\
+            cs=bmap.contourf(self.lons,self.lats,self.var,nlevel,
                     colors='none',hatches=[None,self.method.hatch],
                     alpha=0.,
                     transform=trans)
 
         elif self.method.method=='gis':
-            cs=bmap.arcgisimage(service='ESRI_Imagery_World_2D',xpixels=self.method.xpixels,
+            cs=bmap.arcgisimage(service='ESRI_Imagery_World_2D',
+                    xpixels=self.method.xpixels,
                     dpi=self.method.dpi,verbose=self.method.verbose,
                     transform=trans)
 
[…]
