### 2020-08-28 6e9dc53
doc change regarding basemap->cartopy
README.md:

    matplotlib and cartopy dependency change

Detect-ARs.rst, Track-ARs.rst, conf.py, index.rst, plot.rst:

    new API using cartopy

environment_py2.py, environment_py3.py:

    unpin matplotlib, basemap->cartopy
