+++ docs/conf.py
@@ -64,4 +64,4 @@ napoleon_use_rtype = False
 napoleon_include_init_with_doc = True
 
 autodoc_mock_imports = ['numpy', 'scipy', 'netCDF4', 'matplotlib',
-        'pandas', 'networkx', 'skimage', 'basemap', 'mpl_toolkits']
+        'pandas', 'networkx', 'skimage', 'cartopy', 'mpl_toolkits']
