+++ setup.py
@@ -6,7 +6,7 @@ from setuptools import find_packages, setup
 #from distutils.core import setup
 
 setup(name='ipart',
-        version='3.0.3',
+        version='3.0.4',
         description='IPART is a Python package for the detection and tracking of atmospheric rivers from gridded IVT data using image-processing techniques.',
         author='Guangzhi XU',
         author_email='xugzhi1987@gmail.com',
@@ -30,8 +30,8 @@ setup(name='ipart',
             "pandas",
             "scikit-image",
             "networkx",
-            "matplotlib==2.2.3",
-            "basemap"
+            "matplotlib",
+            "cartopy"
         ],
         packages=find_packages(include=['ipart', 'ipart.*']),
         license='GPL-3.0-or-later'
