### 2020-08-28 ihesp/IPART#5 @Xunius
@kbarnhart @rabernat I've made the transition from basemap to cartopy in these commits:
[commit 1](f0b1572cca5bc6960f7d8607dc905bf234e47f12)
[commit 2](ef281c506d194cc99a413ce9cff60f30e760c41e)
[commit 3](6e9dc53efbbd76b697050e4bf91cab0d4a3d4be7)
[commit 4](0f96df40211303bd2941915a105bb7435e451f0b)

New conda-recipe is updated [here](https://github.com/conda-forge/ipart-feedstock/pull/3)

A side note regarding the cartopy replacement, I feel that we are pushing the deprecation too hard, last time I checked out cartopy it didn't support the latitude/longitude labeling in projections other than PlateCarree and Mercator, this functionality is added in the recent 0.18.0 release in May. It took me a good while to figure out how to plot the data across the dateline, and you can see some relevant issues still open. And basemap is perfectly installable inside conda. Anyway, please let me know whether this issue has been addressed.
