### 2020-08-28 ihesp/IPART#5 @kbarnhart
Thanks for this update @Xunius.

I'll let @rabernat weight in on whether his concerns have been addressed.
