### 2020-09-01 ihesp/IPART#5 @rabernat
Hi @Xunius -- sorry for my long silence. Thanks for all the work you've put into this. It's great that the package is now conda installable!

I don't have strong feelings about basemap vs. cartopy, other than pointing out that basemap is deprecated. Relying on basemap will make it harder to maintain your package in the future. But my most relevant comment was:

> ... more generally, why does your package need to depend on a mapping tool at all? Or any other visualization package? Of course one needs visualization to plot the results, but that's not the same thing as a dependency.

It don't think this question has been addressed. I don't fundamentally understand why you need cartopy / matplotlib *as a dependency*. This is not a visualization package, is it? I see you have some utility routines (e.g. `plotAR`) which call those packages. But these could be made optional? Matplotlib and cartopy are huge dependencies. Imagine that someone wants to use your package as part of an automated system--they don't need any visualization, they just want to detect the ARs and write the data to a file. In its current form, you are *requiring them* to install matplotlib / cartopy in their environment. Your [own tests](https://github.com/ihesp/IPART/tree/master/tests) don't even cover the `plotAR` function, so clearly it is not crucial to the package's functionality. (Conversely, if it is, it should be covered by tests, but that's a separate issue.) You could remove the matplotlib imports from your test suite, as the package is not used at all in the tests.

Regarding the original issue here, installation instructions, I see you have [made some updates](https://ipart.readthedocs.io/en/latest/). However, your recommended installation is still:

> Users are advised to obtain the code of this pacakge from the github page, and create a new conda environment using the environment files provided:

Considering these recommendations differ from 99% of existing python packages, I would like to understand the justification for them. Why is iHESP so special that it requires its own dedicated environment + installation from github? Why can't the installation instructions just simply be?

```
conda install -c conda-forge ipart
```

Wouldn't that be simpler and easier for your users?
