### 2020-09-01 #2407 @kbarnhart
@sadielbartholomew and @rabernat thanks for your review comments thus far. I wanted to check in and see what your time-frame is for revisiting and potentially completing your reviews. As best as I can tell @Xunius has addressed the current outstanding issues that this review has raised.

* ihesp/IPART/issues/5 is open and must be closed, but I think the issues have been addressed.
* ihesp/IPART/issues/9 has been identified as "nice but not necessary"
