### 2020-09-01 ihesp/IPART#6 @rabernat
Great, this is definitely a big improvement. I just have one recommendation. When you reference the project's GitHub page and issue tracker in the docs, provide a hyperlink so it's easier for users to find them.

> I copied from https://github.com/neurohackademy/nh2020-curriculum/blob/master/tu-python-libraries-rokem/03-setup-file.md

If you are copying directly from another package, you will need to provide proper attribution. Unfortunately the licensing on the neurohackacademy repo is not clear (see https://github.com/neurohackademy/nh2020-curriculum/issues/42).
