### 2020-09-01 #2407 @kbarnhart
@rabernat thanks for the update and those comments on the relevant issue threads.

JOSS is trying to be mindful of the impact of COVID on people's availability (see [statement of reduced service mode](https://github.com/openjournals/joss-reviews/issues/2407#issue-647682667) above). Which is to say that we understand things may take longer at the moment.
