### 2020-09-01 #2407 @sadielbartholomew
Thanks to everyone for their patience as well with respect to my review. I am hoping to be able to finish my review tonight, if not by the mid-week at latest.
