### 2020-09-01 ihesp/IPART#5 @Xunius
@rabernat Thanks for coming back.

I used basemap previously for plotting, but also for some computations as well. I used its coordinates transformation function to compute some polygon areas measured in km^2. But that relevant functionality was turned off in a later change. So it is now solely for plotting. Matplotlib's contour function is also used to obtain the contour coordinates, without actually plotting them out. That's why I have to import matplotlib in the test scripts, and change the backend to 'Agg' immediately.

My understanding is that since JOSS requires the authors to provide some notebooks to help the users to get to know the software, plotting modules are therefore required dependencies. Similarly for the "advised" installation method, I feel that it's encouraged to let the user get the git repository, so that they can obtain the notebooks. Besides, the `conda install -c conda-forge ipart` guide is provided just below that, I don't think that's too big a confusion to the readers.
