### 2020-09-01 ihesp/IPART#5 @kbarnhart
I'll chime in quickly with some JOSS-specific guidance: JOSS does not require notebooks, specifically. However, we do require that there is sufficient narrative documentation such that a user can learn to use the software, and API documentation. Often for python packages, Jupyter Notebooks are one of the easiest ways to provide narrative documentation and examples.

However, this is beside the point of whether matplotlib and cartopy are appropriate as dependencies.
