### 2020-09-01 #2407 @sadielbartholomew
(I'm commenting regarding review points out of order with respect to the review list here, purely to address the 'low hanging fruit' firstly.)

@Xunius with regards to the 'Quality of writing' point of the paper itself, I think it is very well written and clear, so well done for that. It is all great as far as I am concerned.

Before I tick that off, though, I wanted to check if you would agree that it would be better to break up the initial 'Summary' section, for the reasons following? I think breaking it into two separate sections would be useful to readers because currently there are four long paragraphs (plus one short one) there, and ideally a summary is shorter and more concise. But also it strikes me that the first two paragraphs are solely about the background , and only the final three discuss IPART itself. So to me there is a natural partition.

Therefore, I would suggest the first two paragraphs can be moved under a section called 'Background' and the rest can stay as the 'Summary'. However, looking at other JOSS papers published I see that plenty have long summary sections, so it is not a blocker to acceptance, by any means. Maybe others prefer one longer section as it flows better, for example. It is of course your paper so if you prefer to keep the whole summary text together as-is, that is fine. What do you think?

(And if anyone else has thoughts on the above, it would be great to hear them.)
