### 2020-09-02 ihesp/IPART#5 @Xunius
@kbarnhart So it is ok to remove cartopy as a dependency but leave the notebooks there? Just put a sentences somewhere in the notebook to tell the user to install cartopy themselves?
