### 2020-09-02 afaab1a
make cartopy optional
README.md, docs/index.rst:

    make cartopy optional.

docs/detect_ARs.py, detect_ARs_generator_version.py, plot.rst:

    reflect the changed plotAR(), plotARTrack() API change.

environment_py2.yml, environment_py3.yml:

    remove cartopy

AR_detector.py:

    move plotAR() into utils.plot

AR_tracer.py:

    move plotAR() into utils.plot and rename to plotARTrack()

utils.plot.py:

    add plotAR(), add plotARTrack().

Index.ipynb, 2_compute_THR.ipynb, 3_detect_ARs.ipynb, 4_track_ARs.ipynb:

    use new plotAR(), plotARTrack().
    Tell user to install cartopy.

scri
[…]
