+++ ipart/utils/plot.py
@@ -1903,3 +1903,219 @@ def plot2(var, method, ax, legend='global', xarray=None, yarray=None,
     cs=plotobj.plot()
 
     return plotobj
+
+
+def plotAR(ardf, ax, lonax):
+    '''Helper function to plot the regions and axes of ARs
+
+    Args:
+        ardf (pandas.DataFrame): table containing AR records.
+        ax (matplotlib axis): axis to plot onto.
+        lonax (ndarray): 1d array of the longitude axis the plot is using.
+    '''
+
+    for ii in range(len(ardf)):
+
+        vv=ardf.iloc[ii]
+        isrelaxkk=vv['is_relaxed']
+
+        # plot contour
+        px=vv['contour_x']
+        py=vv['contour_y']
+
+        #px_segs, py_segs=functions.breakCurveAtEdge(px, py, bmap.llcrnrx,
+        #        bmap.urcrnrx)
+        px_segs, py_segs=functions.breakCurveAtEdge(px, py,
+                #ax.get_extent()[0],
+                #ax.get_extent()[1])
+                lonax[0],
+                lonax[-1])
+        # note that the GeoAxes (of cartopy) doesn't seem to carry these info.
+
+        for xjj, yjj in zip(px_segs, py_segs):
+
+            #xjj,yjj=bmap(xjj,yjj)
+            linewidth=2.0 if isrelaxkk else 2.0
+            linestyle=':' if isrelaxkk else '-'
+            ax.plot(xjj,yjj,color='k',linestyle=linestyle,linewidth=linewidth,
+                    transform=ccrs.PlateCarree())
+
+        # plot axis
+        px=vv['axis_x']
+        py=vv['axis_y']
+
+        #px_segs, py_segs=functions.breakCurveAtEdge(px, py, bmap.llcrnrx,
+        #        bmap.urcrnrx)
+        px_segs, py_segs=functions.breakCurveAtEdge(px, py,
+                #ax.get_extent()[0],
+                #ax.get_extent()[1])
+                lonax[0],
+                lonax[-1])
+        for xjj, yjj in zip(px_segs, py_segs):
+
+            #xjj,yjj=bmap(xjj,yjj)
+            ax.plot(xjj,yjj,'g:',linewidth=2.0,
+                    transform=ccrs.PlateCarree())
+
+        # plot cross flux text
+        '''
+        lenkk=vv['length']
+        areakk=vv['area']
+        widthkk=vv['width']
+        cx=float(vv['centroid_x'])%360
+        cy=float(vv['centroid_y'])
+        cx,cy=bmap(cx,cy)
+
+        strkk=r'ID=%d, $R=%.0f$' %(ii+1,np.sqrt(areakk/3.14)) + '\n'+\
+                r'$L = %d km$' %lenkk +'\n'+\
+                r'$W = %d km$' %widthkk
+
+        ax.annotate(strkk,xy=(cx,cy),
+                horizontalalignment='center',
+                verticalalignment='center',
+                fontsize=8,
+                bbox=dict(facecolor='white',alpha=0.5))
+        '''
+
+    return
+
+
+def plotARTrack(arlist,latax,lonax,ax,full=False,label=None,linestyle='solid',
+        marker=None):
+    '''Plot AR tracks
+
+    Args:
+        arlist (list): list of AR objects to plot.
+        latax,lonax (ndarray): 1darrays giving latitude- and longitude-
+            coordinates of the plotting domain.
+        ax (matplotlib.axis): axis to plot onto.
+
+    Keyword Args:
+        full (bool): if True, plot tracks of an AR from its entire lifecycle.
+                     if False, plot only the track of the last time step.
[…]
