+++ ipart/AR_detector.py
@@ -12,7 +12,7 @@ from skimage import measure
 from skimage import morphology
 from scipy import ndimage
 import matplotlib.pyplot as plt
-import cartopy.crs as ccrs
+#import cartopy.crs as ccrs
 from netCDF4 import date2num
 
 from ipart.utils import rdp
@@ -1075,80 +1075,6 @@ def save2DF(result_dict):
 
     return result_df
 
-
-def plotAR(ardf, ax, lonax):
-    '''Helper function to plot the regions and axes of ARs
-
-    Args:
-        ardf (pandas.DataFrame): table containing AR records.
-        ax (matplotlib axis): axis to plot onto.
-        lonax (ndarray): 1d array of the longitude axis the plot is using.
-    '''
-
-    for ii in range(len(ardf)):
-
-        vv=ardf.iloc[ii]
-        isrelaxkk=vv['is_relaxed']
-
-        # plot contour
-        px=vv['contour_x']
-        py=vv['contour_y']
-
-        #px_segs, py_segs=funcs.breakCurveAtEdge(px, py, bmap.llcrnrx,
-        #        bmap.urcrnrx)
-        px_segs, py_segs=funcs.breakCurveAtEdge(px, py,
-                #ax.get_extent()[0],
-                #ax.get_extent()[1])
-                lonax[0],
-                lonax[-1])
-        # note that the GeoAxes (of cartopy) doesn't seem to carry these info.
-
-        for xjj, yjj in zip(px_segs, py_segs):
-
-            #xjj,yjj=bmap(xjj,yjj)
-            linewidth=2.0 if isrelaxkk else 2.0
-            linestyle=':' if isrelaxkk else '-'
-            ax.plot(xjj,yjj,color='k',linestyle=linestyle,linewidth=linewidth,
-                    transform=ccrs.PlateCarree())
-
-        # plot axis
-        px=vv['axis_x']
-        py=vv['axis_y']
-
-        #px_segs, py_segs=funcs.breakCurveAtEdge(px, py, bmap.llcrnrx,
-        #        bmap.urcrnrx)
-        px_segs, py_segs=funcs.breakCurveAtEdge(px, py,
-                #ax.get_extent()[0],
-                #ax.get_extent()[1])
-                lonax[0],
-                lonax[-1])
-        for xjj, yjj in zip(px_segs, py_segs):
-
-            #xjj,yjj=bmap(xjj,yjj)
-            ax.plot(xjj,yjj,'g:',linewidth=2.0,
-                    transform=ccrs.PlateCarree())
-
-        # plot cross flux text
-        '''
-        lenkk=vv['length']
-        areakk=vv['area']
-        widthkk=vv['width']
-        cx=float(vv['centroid_x'])%360
-        cy=float(vv['centroid_y'])
-        cx,cy=bmap(cx,cy)
-
-        strkk=r'ID=%d, $R=%.0f$' %(ii+1,np.sqrt(areakk/3.14)) + '\n'+\
-                r'$L = %d km$' %lenkk +'\n'+\
-                r'$W = %d km$' %widthkk
-
-        ax.annotate(strkk,xy=(cx,cy),
-                horizontalalignment='center',
-                verticalalignment='center',
-                fontsize=8,
-                bbox=dict(facecolor='white',alpha=0.5))
-        '''
-
-    return
 
 
 def getNormalVectors(point_list, idx):
