+++ scripts/trace_ARs.py
@@ -62,9 +62,9 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 import cartopy.crs as ccrs
+from ipart.utils import plot
 
-from ipart.AR_tracer import readCSVRecord, trackARs, filterTracks, \
-        plotAR
+from ipart.AR_tracer import readCSVRecord, trackARs, filterTracks
 
 
 
@@ -108,7 +108,7 @@ if __name__=='__main__':
         if PLOT:
             figure=plt.figure(figsize=(12,6),dpi=100)
             ax=figure.add_subplot(111, projection=ccrs.PlateCarree())
-            plotAR(tii,latax,lonax,ax,True)
+            plot.plotARTrack(tii,latax,lonax,ax,True)
 
             #----------------- Save plot------------
             plot_save_name='ar_track_%s' %trackidii

[…]
