### 2020-09-02 ihesp/IPART#5 @kbarnhart
@Xunius That is not how I would interpret my comment. My comment was meant to point out that jupyter notebooks are not an explicit requirement of JOSS. However, having narrative documentation and examples are a requirement (and these are commonly created as jupyter notebooks).

One approach that I have seen to communicate install instructions when there are packages needed to run examples but not needed for the package itself is as follows:
- list the core dependencies for the package in the setup.py and conda recipe.
- create a conda environment for running examples that includes the package itself and anything else needed for running the examples (packages listed are often use for visualization).
- create a conda environment for source code development which includes the packages dependencies, and developer tools such as pytest, coveralls, and sphinx.

A nice byproduct of this approach is that one of the "examples" conda environment can be used to specify Binder.

It supports a pretty clean set of install instructions, namely:

1. To use the package install via conda install -c conda-forge ipart
2. To create a conda environment in which the examples can run create an environment with "environment.yml".
3. To create a conda environment for source code modification use "environment-dev.yml"

@rabernat may have comments on this approach.

However, it appears to me that there are still some comments made by @rabernat [above](https://github.com/ihesp/IPART/issues/5#issuecomment-684928842) to address. Specifically (and I paraphrase):

- [ ] Why are the installation instructions different than 99% of python packages
- [ ] Why are matplotlib and cartopy dependencies? (@Xunius started to address this [here](https://github.com/ihesp/IPART/issues/5#issuecomment-684947414)).
