### 2020-09-02 ihesp/IPART#5 @rabernat
This is a really nice summary @kbarnhart.

I think this is a technical issue, related to what it means to create a package vs. documentation / examples of that package. Let me use my experience as an xarray developer to help clarify.

## Dependencies

### Required Dependencies

Required dependencies should be the _bare minimum_ of other packages required by your package in order to run its core functionality. Xarray's dependencies, as specified in its setup.cfg, are just
```
    numpy >= 1.15
    pandas >= 0.25
    setuptools >= 38.4
```
https://github.com/pydata/xarray/blob/2acd0fc6563c3ad57f16e6ee804d592969419938/setup.cfg#L76-L82

These _must_ be installed for xarray to run, and if you try to install xarray with pip or conda, these dependencies will automatically be installed as well. Setuptools handles that for you.

So you should ask yourself: _what are the bare minimum requirements to run iHESP's core functionality?_

### Optional Dependencies

Optional dependencies are extra packages which, if installed, enable extra optional features in your package. For xarray, matlplotlib in an optional dependency. It is not required, but some functions use it (like all of the plotting functions). If you try to use these functions without matplotlib installed, you just get an error when xarray tries to `import matplotlib`. But the rest of the package works fine.

## Environments

An environment is not the same as the dependencies. An environment is a self-contained python installation. Most packages define several different environments for different purposes, _but users should not be required or expected to use these exact environments._ Most users already have existing environments that they know and like, and they don't want to create a new environment just for your package.

### Test environments

In xarray, we define 8 different environments for testing:
https://github.com/pydata/xarray/tree/master/ci/requirements
These environments contain different combinations of python and required / optional dependencies. At least one environment should have _all_ of the required dependencies, to make sure you test _all_ of your package's code. These environments are built every time our CI runs.

### Documentation / Demo Environments

Xarray's documentation has it's own environment:
https://github.com/pydata/xarray/blob/master/ci/requirements/doc.yml
This environment contains many optional dependencies, as well as other packages unrelated to xarray itself, like sphinx for actually building the docs.

Finally, we have a binder environment, designed to let users try xarray in the cloud without installing anything:
https://github.com/pydata/xarray/blob/master/.binder/environment.yml
It includes many optional dependencies. The binder links appear in our documentation, like on [this page](http://xarray.pydata.org/en/stable/examples/monthly-means.html).

---

This is a lot to digest. If this is your first time putting together a python package, it may be overwhelming. My point is that looking at the best practices established by mature, widely-used python packages can be a very useful guide.
