### 2020-09-02 a482843
installation guide change
README.md, index.rst:

    installation guide change.

environment_py2.py, environment_py3.py:

    add cartopy to list
