+++ environment_py2.yml
@@ -11,4 +11,5 @@ dependencies:
   - scikit-image
   - pandas
   - networkx
+  - cartopy
 
