### 2020-09-06 f11e0e2
minor wording change in index.rst
