### 2020-09-06 c81152d
remove cartopy from test_install.py, version to 3.0.6
