### 2020-09-06 ihesp/IPART#5 @Xunius
@kbarnhart @rabernat Thanks for all these information. Now I understand your point better.

So I have tried to make cartopy an optional dependency and make it installable via the environment file approach, in these commits:

1. [make cartopy optional](afaab1ac645cfbf5b1a9ef023cae624657b1ab95)
2. [installation guide change](a482843efa644ba4f684a7d97ebf5660dabe5e11)
3. [minor wording change in index.rst](f11e0e2d871da2b1886ca149d9473dcec242dfca)
4. [remove cartopy from test_install.py](c81152d2466c92a7a856788d0bc88d78520d6d56).

and the updated [conda recipe](https://github.com/conda-forge/ipart-feedstock/pull/4).

Regarding matplotlib, I'm still using it to compute contour coordinates, so it's not only for plotting, therefore a required dependency.
