### 2020-09-08 ihesp/IPART#5 @rabernat
> Regarding matplotlib, I'm still using it to compute contour coordinates, so it's not only for plotting, therefore a required dependency.

Makes sense. 👍

> * [make cartopy optional](afaab1ac645cfbf5b1a9ef023cae624657b1ab95)
> * [installation guide change](a482843efa644ba4f684a7d97ebf5660dabe5e11)
> * [minor wording change in index.rst](f11e0e2d871da2b1886ca149d9473dcec242dfca)
> * [remove cartopy from test_install.py](c81152d2466c92a7a856788d0bc88d78520d6d56).

None of these links works for me, so it's hard for me to directly validate your claims / approve your changes.
