### 2020-09-08 ihesp/IPART#5 @Xunius
@rabernat My bad. I copied the wrong string. Please check the updated links below:

* [make cartopy optional](https://github.com/ihesp/IPART/commit/afaab1ac645cfbf5b1a9ef023cae624657b1ab95)
* [installation guide change](https://github.com/ihesp/IPART/commit/a482843efa644ba4f684a7d97ebf5660dabe5e11)
* [minor wording change in index.rst](https://github.com/ihesp/IPART/commit/f11e0e2d871da2b1886ca149d9473dcec242dfca)
* [remove cartopy from test_install.py](https://github.com/ihesp/IPART/commit/c81152d2466c92a7a856788d0bc88d78520d6d56)
