### 2020-09-15 ihesp/IPART#5 @kbarnhart
@Xunius with one small exception I think these changes look good. Of course, I'll let @rabernat chime in and indicate what he thinks.

My only recommendation would be to put the cartopy import in ipart\utils\plot.py in a try except block with an informative error message telling a user that they've tried to use a part of ipart which relies on the optional dependency. For example:

```python
try:
    import cartopy
except ImportError:
    raise ImportError("Your message here")
```
