### 2020-09-15 #2407 @kbarnhart
Thanks for these comments @sadielbartholomew. I think that organizing the beginning of the paper either way would be fine. I generally tend towards using descriptive section titles, so I'd lean slightly towards adding them.
