### 2020-09-24 ab58c86
use kwarg in sliceData() in compute_thr_singlefile.py
