### 2020-09-26 1523b42
remove project kwarg from plot2() in compute_ivt.py
