### 2020-09-30 ihesp/IPART#6 @kbarnhart
@Xunius it looks like @rabernat's issue (neurohackademy/nh2020-curriculum#42) has clarified the license on the neurohackacademy. Once you have provided appropriate attribution for the contribution guidelines I think this issue can be closed.
