### 2020-09-30 #2407 @kbarnhart
@Xunius I wanted to check in and see if you had any thoughts regarding @sadielbartholomew's comment about paper structure.

At present there are two remaining issues in the source repository associated with this review. I've just commented there to move us forward. Once those are resolved I suspect it would be appropriate for @sadielbartholomew and @rabernat to re-assess the review checklist and identify if there are any remaining issues or if the submission is ready to move forward.

As always, let me know if you have any questions or comments.
