### 2020-10-09 84e5fb9
add "background" section to paper.md
