### 2020-10-09 #2407 @Xunius
@sadielbartholomew @kbarnhart I've just updated the JOSS.md structure to add a opening "Background" section as suggested.
