### 2020-10-09 ad7fd57
add attribution to Neurohackademy 2020 curriculum
Addressing the issue https://github.com/ihesp/IPART/issues/6#issuecomment-701461940.
