### 2020-10-09 ihesp/IPART#6 @Xunius
@rabernat @kbarnhart. Hi all, I've just added attribution to neurohackacademy in [this commit](https://github.com/ihesp/IPART/commit/ad7fd5708edf799084ef44f5abdfbad93ba25850).
