### 2020-10-10 0dfeaeb
add reference to GMD in joss paper and README
