+++ joss/paper.bib
@@ -673,3 +673,15 @@ ARTICLE-NUMBER = {628},
 ISSN = {2073-4433},
 DOI = {10.3390/atmos11060628}
 }
+
+@Article{Xu2020b,
+AUTHOR = {Xu, G. and Ma, X. and Chang, P. and Wang, L.},
+TITLE = {Image-processing-based atmospheric river tracking  method version 1 ({IPART}-1)},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {13},
+YEAR = {2020},
+NUMBER = {10},
+PAGES = {4639--4662},
+URL = {https://gmd.copernicus.org/articles/13/4639/2020/},
+DOI = {10.5194/gmd-13-4639-2020}
+}
