### 2020-10-10 #2407 @Xunius
@rabernat @kbarnhart @sadielbartholomew Thanks for all your efforts put into this work.
I'm writing also to inform you that I'm making some minor edits to the joss paper and the README by adding the the reference to the paper on this method that just got published a few days ago. The changes are in [this commit](https://github.com/ihesp/IPART/commit/0dfeaeb0d3f06aa45a6d2427bbe7611f1286077c).
