### 2020-10-13 7b3d59d
fix AR_tracer.py bug in counting nonrelaxed ARs
IMPORTANT fix: the number of non-relaxed points in an AR track was
counted wrong. It should be .count(False) instead of .count(True)
