### 2020-10-14 #2407 @kbarnhart
@Xunius thank you for the update. I wanted to ask if you could clarify the difference between this recently published paper in GMD and the manuscript considered here?
