### 2020-10-15 #2407 @Xunius
@kbarnhart The GMD one contains the full derivation of the methods, parameter sensitivity tests, some comparisons with existing methods in the form of case studies, and lots of discussions on the implications on climate projection applications etc.. While the JOSS manuscript is providing just enough information to support the Python implementation of the algorithms. The GMD paper was actually submitted earlier than JOSS, it also took a long time for the review.
