### 2020-10-19 ihesp/IPART#12 @sadielbartholomew
ImportError ('funcs' from 'ipart.utils') impeding functionality
Towards openjournals/joss-reviews#2407 review, I am trying to confirm a working installation and sufficient claimed functionality but, after a standard install as per the current instructions, I am running immediately into the same exception, `ImportError`.

Namely, using the following commands in the root directory of my fork of the `IPART` repo, updated to the current state of the IPART `master` branch:

```bash
$ conda env create -f environment_py3.yml
...
$ conda activate ipartpy3
(ipartpy3) $ pip install -e .
...
(ipartpy3) $ python
...
(ipartpy3) $ python -m unittest discover -s tests
... <full output copied below>
```
the unit tests, the scripts and the notebooks all immediately error from various usage of `from ipart.utils import funcs` giving the ultimate exception of:

```pytb
ImportError: cannot import name 'funcs' from 'ipart.utils' (unknown location)
```

(although running `import ipart` in the Python session is successful and does not output that, or indeed any, error). I have included the full unit test run, and some representative script, outputs below if they may be helpful.

It seems an old location may be referenced for the import in question, or perhaps it is an attempted circular import. Whatever the cause, this issue needs resolving since this error is preventing any access to most of the functionality of IPART. Given the form it should be quick to fix.

### Exact outputs

All with the same underlying error, but just to illustrate the extent to which it is emerging...

#### Some script outputs

```console
$ cd scripts
$ python compute_ivt.py
Traceback (most recent call last):
  File "compute_ivt.py", line 28, in <module>
    from ipart.utils import funcs
$ python trace_ARs.py
Traceback (most recent call last):
  File "trace_ARs.py", line 65, in <module>
    from ipart.utils import plot
  File "/home/sadie/IPART/ipart/utils/plot.py", line 11, in <module>
    from . import funcs as functions
ImportError: cannot import name 'funcs' from 'ipart.utils' (unknown location)

```
#### Full test output
Running `python -m unittest discover -s tests` under the root `IPART` repo directory returns:

```pytb
EEEEE
ERROR: test_data (unittest.loader._FailedTest)
ImportError: Failed to import test module: test_data
Traceback (most recent call last):
  File "/home/sadie/anaconda3/envs/ipartpy3/lib/python3.7/unittest/loader.py", line 436, in _find_test_path
    module = self._get_module_from_name(name)
  File "/home/sadie/anaconda3/envs/ipartpy3/lib/python3.7/unittest/loader.py", line 377, in _get_module_from_name
    __import__(name)
  File "/home/sadie/IPART/tests/test_data.py", line 15, in <module>
    from ipart.utils import funcs
ImportError: cannot import name 'funcs' from 'ipart.utils' (unknown location)

ERROR: test_ipart (test_install.TestIPRT)
Traceback (most recent call last):
  File "/home/sadie/IPART/tests/test_install.py", line 13, in test_ipart
    from ipart.thr import THR, rotatingTHR
ModuleNotFoundError: No module named 'ipart.thr'

ERROR: test_river_tracker1 (unittest.loader._FailedTest)
ImportError: Failed to import test module: test_river_tracker1
Traceback (most recent call last):
  File "/home/sadie/anaconda3/envs/ipartpy3/lib/python3.7/unittest/loader.py", line 436, in _find_test_path
    module = self._get_module_from_name(name)
  File "/home/sadie/anaconda3/envs/ipartpy3/lib/python3.7/unittest/loader.py", line 377, in _get_module_from_name
    __import__(name)
  File "/home/sadie/IPART/tests/test_river_tracker1.py", line 14, in <module>
    from ipart.utils import funcs
ImportError: cannot import name 'funcs' from 'ipart.utils' (unknown location)

ERROR: test_river_tracker2 (unittest.loader._FailedTest)
ImportError: Failed to import test module: test_river_tracker2
Traceback (most recent call last):
  File "/home/sadie/anaconda3/envs/ipartpy3/lib/python3.7/unittest/loader.py", line 436, in _find_test_path
    module = self._get_module_from_name(name)
  File "/home/sadie/anaconda3/envs/ipartpy3/lib/python3.7/unittest/loader.py", line 377, in _get_module_from_name
    __import__(name)
  File "/home/sadie/IPART/tests/test_river_tracker2.py", line 13, in <module>
    from ipart.AR_tracer import readCSVRecord, trackARs, filterTracks
  File "/home/sadie/IPART/ipart/AR_tracer.py", line 19, in <module>
    from ipart.utils import funcs
ImportError: cannot import name 'funcs' from 'ipart.utils' (unknown location)

ERROR: test_thr (unittest.loader._FailedTest)
ImportError: Failed to import test module: test_thr
Traceback (most recent call last):
  File "/home/sadie/anaconda3/envs/ipartpy3/lib/python3.7/unittest/loader.py", line 436, in _find_test_path
    module = self._get_module_from_name(name)
  File "/home/sadie/anaconda3/envs/ipartpy3/lib/python3.7/unittest/loader.py", line 377, in _get_module_from_name
    __import__(name)
  File "/home/sadie/IPART/tests/test_thr.py", line 14, in <module>
    from ipart.utils import funcs
ImportError: cannot import name 'funcs' from 'ipart.utils' (unknown location)

Ran 5 tests in 0.001s

FAILED (errors=5)
```
