### 2020-10-19 ihesp/IPART#12 @Xunius
@sadielbartholomew Thanks for the report.
I repeated your steps but didn't manage to reproduce the error. I didn't get the same `ImportError`, and the unit tests ran OK (aside from another unrelated error which I have fixed).

I searched a bit about this `unknown location` error, I don't think I got anything particularly useful. Some one said it might be a name conflict. Do you happen to have something named `funcs` that is importable, maybe related to one of the records in the `PYTHONPATH` env variable?
