### 2020-10-19 de215e6
fix unit test after fixing the track filtering bug
Related to this commit: 7b3d59d86b5b0fc4e7e6a2f15b434219e3310558
