### 2020-10-19 ihesp/IPART#12 @sadielbartholomew
Hi @Xunius, thanks for your quick response and investigations. If this is indeed a mistake on my part, which seems the likely case, then I am sorry. However, I did create a clean conda environment as shown in my opening comment, so I want to make sure that it is not an issue others an run into before I close this. Please bear with me while I investigate.

> Some one said it might be a name conflict. Do you happen to have something named funcs that is importable, maybe related to one of the records in the PYTHONPATH env variable?

Good idea and that is probably the case. I tried running a script with `python -v` to get further details about the imports, and though I cannot find any other `funcs` mentioned in the verbose output, I do see there is a potential namespace clash on `utils` as there is also a `numpy.lib.utils` imported. So I suspect that may be the underlying issue. I am just trying to confirm this or otherwise.

I'll report back by the end of today on my findings. Thanks for your patience.
