### 2020-10-19 #2407 @sadielbartholomew
Hi @Xunius , while I am looking into the issue referenced above which is likely to be a user error on my part, I wanted to make one final small suggestion for an update to the paper itself.

I noticed that, in September and since we began this review, NumPy released a new paper which they now document as their preferred citation reference, [as described on this page](https://numpy.org/citing-numpy/). I think it might be nice to use that up-to-date reference rather than one from 2011. It is of course up to you as to which you go with, though; feel free to keep the old one if you prefer.
