### 2020-10-19 ihesp/IPART#12 @sadielbartholomew
Hi again, I have isolated the point at which the error is introduced. It is the `pip install -e .` command, which the installation instructions in the README state to use but not the docs. Either way, If I instead install IPART via conda, `conda install -c conda-forge ipart`, the unit tests run and pass for me. But with the development install with `pip install -e .` they fail with the `ImportError` as described above.

So this would imply that codebase in the state uploaded to conda-forge does not have produce issue, but the current state of the codebase as in the `master` branch does have the issue present. It seems like a change made since the [conda-forge upload ~2 months ago](https://anaconda.org/conda-forge/ipart) something has changed to result in the error.

Given that for the release of the JOSS paper we will need you to create a tagged release of IPART to associate with the paper and to archive, we need the current branch to be working and to provide the necessary functionality. So in conclusion, I do think this is a problem that needs resolving, because there is something broken on `master`, rather than an issue with my own environment setup.

Does that make sense? I can always ask the editor or my co-reviewer to double check to see if they can see this issue too with `pip install -e .` if you would like someone to double check. Thanks.
