### 2020-10-20 ihesp/IPART#12 @Xunius
@sadielbartholomew Hi. But that doesn't explain why I could not reproduce the error. I can install from conda or `pip install -e .`, both worked for me. The last conda upload was fixing the dependency issue, I removed the `basemap` module from the dependency list, then nothing significant got changed since then. So I don't think I introduced the error to `master` that would trigger such a dramatic failure.
