### 2020-10-20 2539d66
update numpy citation
