### 2020-10-20 ihesp/IPART#12 @sadielbartholomew
OK that's fine, if it also works for you with `pip install -e .` then I am thinking again it is likely to be an error in my environment. I've been creating a clean `ipartpy3` env but maybe there is something in my base env or similar that is interacting. Sorry about this, I will try to get to the bottom of it today. Thanks for bearing with me.
