### 2020-10-20 c51b2a6
version to 3.0.7 for a new conda upload
