### 2020-10-20 ihesp/IPART#12 @Xunius
@sadielbartholomew I'm creating an update on the conda version, but my internet has been real slow today. I'll give you a notice when that's done, then you can maybe give it another try and see if the same issue persists in the conda install as well.
