### 2020-10-20 ihesp/IPART#12 @sadielbartholomew
> then you can maybe give it another try and see if the same issue persists in the conda install as well.

The conda install is fine, there is no such issue when I install with conda, the error just appears with `pip install -e .`. So it is the current codebase that might be the issue, if not something about the way I installed it or my environment.
