### 2020-10-21 ihesp/IPART#12 @Xunius
@sadielbartholomew I've updated the conda version to the newest 3.0.7 so it is now the same as the master branch. And I did a test using conda and `pip install -e .`, again both fine.
Are you familiar with google colab? Maybe we could try the `pip` approach from within google colab? I've never used it before, I didn't managed to run the `conda env create -f env.yml` command, probably something missing. Here is my current attempt: https://colab.research.google.com/drive/1dK5nH9hu9x7UC_tlLq4aHvC73f_lrt35?usp=sharing
