### 2020-10-25 #2407 @sadielbartholomew
Hi @kbarnhart, I have a little time later tonight to have a look. I'll see how far I can get and let you know a new estimate after that.

I was seeing some errors (as per linked Issue above) that are probably related to issues in my environment, but I need to be that is the case before I can proceed. Until I fix those I can't test the functionality etc. properly. So that is what is holding me up.
