### 2020-10-28 ihesp/IPART#12 @sadielbartholomew
Sorry @Xunius, your comment above must have fell off my radar. I have been finishing my review and have since worked out why I was seeing those errors. The good news is that it was indeed a user error on my part, namely that some of the modules had disappeared in my forked repo, including the `func.py`, so naturally the tests were complaining that they couldn't import that module! I have no idea how it happened, perhaps a `rm` command was passed without my realising.

Apologies for the mistake. The errors reported above are not at all issues with the library, rather  that I managed to mess up my repo somehow. I'll close this now  :slightly_smiling_face:. FYI I have ensured my repo is in sync with the `master` branch and with no missing files and all the tests pass:

```python
rootdir: /home/sadie/IPART
collected 12 items

tests/test_data.py ......                                                [ 50%]
tests/test_install.py .                                                  [ 58%]
tests/test_river_tracker1.py ...                                         [ 83%]
tests/test_river_tracker2.py .                                           [ 91%]
tests/test_thr.py .                                                      [100%]

============================= 12 passed in 14.26s ==============================
```
