### 2020-10-29 c9d3e6c
JOSS paper proofread: minor formatting and wording edits
