+++ joss/paper.bib
@@ -156,7 +156,7 @@ URL = {
 
 
 @article{Gimeno2014,
-author = {Gimeno, Luis and Nieto, Raquel and V{\~A}zquez, Marta and Lavers, David A.},
+author = {Gimeno, Luis and Nieto, Raquel and V{\'a}zquez, Marta and Lavers, David A.},
 isbn = {2296-6463},
 year = {2014},
 keywords = {atmospheric branch of the, atmospheric rivers, atmospheric rivers, transport of moisture, atmosph, extratropical cyclones, hydrological cycle, intense, precipitation, transport of moisture},
