### 2020-10-29 #2407 @sadielbartholomew
I am very happy to now recommend this paper for publication, with many congratulations to Guangzhi and thanks for their patience in waiting for the final review. IPART is clearly an innovative package and I expect it will be a very useful tool for research into or involving atmospheric rivers.

I do have a small number of final comments/questions about the paper content that @Xunius may or may not wish to make some last amendments based upon. Either way, it is not a blocker to my acceptance:

* at one point in the paper you state:

  > "under a warming climate as the atmospheric moisture level is expected to increase"

  I take it this is a well-known and well-grounded statement (I personally don't know how established that is as a fact)? If it is not, or not so much, I advise you consider citing something against it for validation or further detail, etc.

* at another point there is a general statement that is worded in a way which I confident is not strictly technically correct, namely:

  > using the Python programming language as a wrapper to some well-established numeric packages including `numpy`, `scikit-image` and `networkx` etc.

  I don't think it is right to call Python a 'wrapper' to Python modules. If anything it would be vice versa with e.g. `numpy` "wrapping" (in the sense of extending) Python, but that is also not really true I would say. So I suggest that you re-word that sentence. Something like "Python programming language *with* some well-established ..." sounds intuitive to me.
* I notice that in Figure 1 of the paper, all of the sub-plots have an identical and identically-scaled colourbar. In the interest of making the full plot easier to interpret and the sub-plots easier to compare, I recommend hiding all but the lowest colourbar if that is simple enough to do with your plotting code.

Lastly, I have also put a small number of Issues up and PRs in to address some minor aspects I have noticed, but other than #14 which makes some suggestions for minor formatting and wording amendments to the paper after my final read-through, none of these need to be addressed before the paper is published. I am leaving them as notes on potential future work.

Thanks all.
