### 2020-10-29 #2407 @sadielbartholomew
(That, as linked to this Issue, is now all of the PRs and Issues I wanted to raise, FYI. To recap, only ihesp/IPART#14 needs to be looked at before in my eyes the paper can be published; the rest are suggestions for further work as general improvements.)
