### 2020-10-29 ihesp/IPART#16 @Xunius
It is not practical to add data to the repo for this type of task. The data people use for detecting such things came in GBs for real world applications. The notebook is just a toy example. So, no, I'm not adding new datasets, and I think the scripts folder is optional to begin with, when installing via conda one doesn't even get the scripts.
