### 2020-10-29 #2407 @kbarnhart
@sadielbartholomew thank you for completing your review! I agree with your distinction between issues should be addressed before publication (https://github.com/ihesp/IPART/pull/14) and others.

@Xunius once you've addressed the comments by @sadielbartholomew above, please ping me here and I'll start the final steps in the JOSS review process.
