### 2020-10-30 f9306f3
copyedits and bib formatting
