### 2020-10-30 #2407 @kbarnhart
@Xunius I've made my final copyedits on ihesp/IPART/pull/17. Please review and merge them.

After merging, please do the following:

- [ ] Make a tagged release of your software, and report the version tag of the archived version here.

I can then move forward with recommending the submission be accepted.
