### 2020-10-31 #2407 @Xunius
@kbarnhart Thanks for the instructions.

* [ihesp/IPART/pull/17](https://github.com/ihesp/IPART/pull/17) is merged.
* A tagged release of the final state after merging the above: https://github.com/ihesp/IPART/releases/tag/v3.0.8
* DOI of an archive in Zenodo: 10.5281/zenodo.4164826.
