### 2020-11-02 #2407 @kbarnhart
@Xunius I've now recommended that this submission be accepted. The JOSS handling editor in chief will manage final publication.

I agree with the reviewers that the package is much improved since submission. Congratulations on all of your hard work.
